Generate rmarkdown codebook

If you pass the object resulting from a call to formr_results to this function, it will generate a markdown codebook for this object.

codebook(results, reliabilities = NULL, survey_repetition = c("auto",
  "single", "repeated_once", "repeated_many"), missingness_report = TRUE,
  indent = "#")

Arguments

results
a formr results table with attributes set on items and scales

reliabilities
a named list with one entry per scale and one or several printable reliability computations for this scale. if NULL, computed on-the-fly using compute_reliabilities

survey_repetition
defaults to "auto" which is to try to determine the level of repetition from the "session" and "created" variables. Other values are: single, repeated_once, repeated_many

missingness_report
whether to print a missingness report. Turn off if this gets too complicated and you need a custom solution (e.g. in case of random missings).

indent
add # to this to make the headings in the components lower-level. defaults to beginning at h2

Examples

# see vignette