Codebook metadata table
Source: R/codebook.R

will generate a table combining metadata from variable attributes with data summaries generated using skimr::skim_to_wide()

codebook_table(results)

Arguments

results
a data frame, ideally with attributes set on variables

Examples

data("bfi")
codebook_table(bfi)

#>             name
#> 1        session
#> 2        created
#> 3       modified
#> 4          ended
#> 5        expired
#> 6    BFIK_open_2
#> 7  BFIK_agree_4R
#> 8   BFIK_extra_2
#> 9  BFIK_agree_1R
#> 10   BFIK_open_1
#> 11 BFIK_neuro_2R
#> 12  BFIK_consc_3
#> 13  BFIK_consc_4
#> 14 BFIK_consc_2R
#> 15 BFIK_agree_3R
#> 16 BFIK_extra_3R
#> 17  BFIK_neuro_3
#> 18  BFIK_neuro_4
#> 19  BFIK_agree_2
#> 20  BFIK_consc_1
#> 21   BFIK_open_4
#> 22  BFIK_extra_4
#> 23 BFIK_extra_1R
#> 24   BFIK_open_3
#> 25    BFIK_agree
#> 26     BFIK_open
#> 27    BFIK_consc
#> 28    BFIK_extra
#> 29    BFIK_neuro
#>                                                                         label
#> 1                                                                        NULL
#> 2                                                    user first opened survey
#> 3                                                     user last edited survey
#> 4                                                        user finished survey
#> 5                                                                        NULL
#> 6                       __Ich bin tiefsinnig, denke gerne über Sachen nach.__
#> 7        __Ich kann mich schroff und abweisend anderen gegenüber verhalten.__
#> 8            __Ich bin begeisterungsfähig und kann andere leicht mitreißen.__
#> 9                                  __Ich neige dazu, andere zu kritisieren.__
#> 10                                       __Ich bin vielseitig interessiert.__
#> 11 __Ich bin entspannt, lasse mich durch Stress nicht aus der Ruhe bringen.__
#> 12                                     __Ich bin tüchtig und arbeite flott.__
#> 13                              __Ich mache Pläne und führe sie auch durch.__
#> 14                                    __Ich bin bequem, neige zur Faulheit.__
#> 15                          __Ich kann mich kalt und distanziert verhalten.__
#> 16                               __Ich bin eher der "stille Typ", wortkarg.__
#> 17                                            __Ich mache mir viele Sorgen.__
#> 18                                  __Ich werde leicht nervös und unsicher.__
#> 19  __Ich schenke anderen leicht Vertrauen, glaube an das Gute im Menschen.__
#> 20                                       __Ich erledige Aufgaben gründlich.__
#> 21                   __Ich schätze künstlerische und ästhetische Eindrücke.__
#> 22                                 __Ich gehe aus mir heraus, bin gesellig.__
#> 23                                __Ich bin eher zurückhaltend, reserviert.__
#> 24             __Ich habe eine aktive Vorstellungskraft, bin phantasievoll.__
#> 25                      4 BFIK_agree items averaged with aggregation_function
#> 26                       4 BFIK_open items averaged with aggregation_function
#> 27                      4 BFIK_consc items averaged with aggregation_function
#> 28                      4 BFIK_extra items averaged with aggregation_function
#> 29                      3 BFIK_neuro items averaged with aggregation_function
#>             type type_options data_type ordered
#> 1           NULL         NULL character    <NA>
#> 2           NULL         NULL   POSIXct    <NA>
#> 3           NULL         NULL   POSIXct    <NA>
#> 4           NULL         NULL   POSIXct    <NA>
#> 5           NULL         NULL   logical    <NA>
#> 6  rating_button            5  labelled   FALSE
#> 7  rating_button            5  labelled   FALSE
#> 8  rating_button            5  labelled   FALSE
#> 9  rating_button            5  labelled   FALSE
#> 10 rating_button            5  labelled   FALSE
#> 11 rating_button            5  labelled   FALSE
#> 12 rating_button            5  labelled   FALSE
#> 13 rating_button            5  labelled   FALSE
#> 14 rating_button            5  labelled   FALSE
#> 15 rating_button            5  labelled   FALSE
#> 16 rating_button            5  labelled   FALSE
#> 17 rating_button            5  labelled   FALSE
#> 18 rating_button            5  labelled   FALSE
#> 19 rating_button            5  labelled   FALSE
#> 20 rating_button            5  labelled   FALSE
#> 21 rating_button            5  labelled   FALSE
#> 22 rating_button            5  labelled   FALSE
#> 23 rating_button            5  labelled   FALSE
#> 24 rating_button            5  labelled   FALSE
#> 25          NULL         NULL   numeric    <NA>
#> 26          NULL         NULL   numeric    <NA>
#> 27          NULL         NULL   numeric    <NA>
#> 28          NULL         NULL   numeric    <NA>
#> 29          NULL         NULL   numeric    <NA>
#>                                                                                                                   value_labels
#> 1                                                                                                                         NULL
#> 2                                                                                                                         NULL
#> 3                                                                                                                         NULL
#> 4                                                                                                                         NULL
#> 5                                                                                                                         NULL
#> 6  1: Trifft überhaupt nicht zu=1,\n2=2,\n3=3,\n4=4,\n5: Trifft voll und ganz zu=5,\nItem was never rendered for this user.=NA
#> 7  1: Trifft überhaupt nicht zu=5,\n2=4,\n3=3,\n4=2,\n5: Trifft voll und ganz zu=1,\nItem was never rendered for this user.=NA
#> 8  1: Trifft überhaupt nicht zu=1,\n2=2,\n3=3,\n4=4,\n5: Trifft voll und ganz zu=5,\nItem was never rendered for this user.=NA
#> 9  1: Trifft überhaupt nicht zu=5,\n2=4,\n3=3,\n4=2,\n5: Trifft voll und ganz zu=1,\nItem was never rendered for this user.=NA
#> 10 1: Trifft überhaupt nicht zu=1,\n2=2,\n3=3,\n4=4,\n5: Trifft voll und ganz zu=5,\nItem was never rendered for this user.=NA
#> 11 1: Trifft überhaupt nicht zu=5,\n2=4,\n3=3,\n4=2,\n5: Trifft voll und ganz zu=1,\nItem was never rendered for this user.=NA
#> 12 1: Trifft überhaupt nicht zu=1,\n2=2,\n3=3,\n4=4,\n5: Trifft voll und ganz zu=5,\nItem was never rendered for this user.=NA
#> 13 1: Trifft überhaupt nicht zu=1,\n2=2,\n3=3,\n4=4,\n5: Trifft voll und ganz zu=5,\nItem was never rendered for this user.=NA
#> 14 1: Trifft überhaupt nicht zu=5,\n2=4,\n3=3,\n4=2,\n5: Trifft voll und ganz zu=1,\nItem was never rendered for this user.=NA
#> 15 1: Trifft überhaupt nicht zu=5,\n2=4,\n3=3,\n4=2,\n5: Trifft voll und ganz zu=1,\nItem was never rendered for this user.=NA
#> 16 1: Trifft überhaupt nicht zu=5,\n2=4,\n3=3,\n4=2,\n5: Trifft voll und ganz zu=1,\nItem was never rendered for this user.=NA
#> 17 1: Trifft überhaupt nicht zu=1,\n2=2,\n3=3,\n4=4,\n5: Trifft voll und ganz zu=5,\nItem was never rendered for this user.=NA
#> 18 1: Trifft überhaupt nicht zu=1,\n2=2,\n3=3,\n4=4,\n5: Trifft voll und ganz zu=5,\nItem was never rendered for this user.=NA
#> 19 1: Trifft überhaupt nicht zu=1,\n2=2,\n3=3,\n4=4,\n5: Trifft voll und ganz zu=5,\nItem was never rendered for this user.=NA
#> 20 1: Trifft überhaupt nicht zu=1,\n2=2,\n3=3,\n4=4,\n5: Trifft voll und ganz zu=5,\nItem was never rendered for this user.=NA
#> 21 1: Trifft überhaupt nicht zu=1,\n2=2,\n3=3,\n4=4,\n5: Trifft voll und ganz zu=5,\nItem was never rendered for this user.=NA
#> 22 1: Trifft überhaupt nicht zu=1,\n2=2,\n3=3,\n4=4,\n5: Trifft voll und ganz zu=5,\nItem was never rendered for this user.=NA
#> 23 1: Trifft überhaupt nicht zu=5,\n2=4,\n3=3,\n4=2,\n5: Trifft voll und ganz zu=1,\nItem was never rendered for this user.=NA
#> 24 1: Trifft überhaupt nicht zu=1,\n2=2,\n3=3,\n4=4,\n5: Trifft voll und ganz zu=5,\nItem was never rendered for this user.=NA
#> 25                                                                                                                        NULL
#> 26                                                                                                                        NULL
#> 27                                                                                                                        NULL
#> 28                                                                                                                        NULL
#> 29                                                                                                                        NULL
#>    optional                                          scale_item_names
#> 1      NULL                                                      NULL
#> 2      NULL                                                      NULL
#> 3      NULL                                                      NULL
#> 4      NULL                                                      NULL
#> 5      NULL                                                      NULL
#> 6         0                                                      NULL
#> 7         0                                                      NULL
#> 8         0                                                      NULL
#> 9         0                                                      NULL
#> 10        0                                                      NULL
#> 11        0                                                      NULL
#> 12        0                                                      NULL
#> 13        0                                                      NULL
#> 14        0                                                      NULL
#> 15        0                                                      NULL
#> 16        0                                                      NULL
#> 17        0                                                      NULL
#> 18        0                                                      NULL
#> 19        0                                                      NULL
#> 20        0                                                      NULL
#> 21        0                                                      NULL
#> 22        0                                                      NULL
#> 23        0                                                      NULL
#> 24        0                                                      NULL
#> 25     NULL BFIK_agree_4R, BFIK_agree_1R, BFIK_agree_3R, BFIK_agree_2
#> 26     NULL        BFIK_open_2, BFIK_open_1, BFIK_open_4, BFIK_open_3
#> 27     NULL   BFIK_consc_3, BFIK_consc_4, BFIK_consc_2R, BFIK_consc_1
#> 28     NULL  BFIK_extra_2, BFIK_extra_3R, BFIK_extra_4, BFIK_extra_1R
#> 29     NULL                 BFIK_neuro_2R, BFIK_neuro_3, BFIK_neuro_4
#>    item_order missing complete  n empty n_unique               top_counts count
#> 1        NULL       0       28 28     0       28                     <NA>  <NA>
#> 2        NULL       0       28 28  <NA>       28                     <NA>  <NA>
#> 3        NULL       0       28 28  <NA>       28                     <NA>  <NA>
#> 4        NULL       0       28 28  <NA>       28                     <NA>  <NA>
#> 5        NULL      28        0 28  <NA>     <NA>                     <NA>    28
#> 6           4       0       28 28  <NA>        4 4: 15, 5: 10, 3: 2, 2: 1  <NA>
#> 7           5       0       28 28  <NA>        5  2: 11, 4: 7, 3: 5, 5: 3  <NA>
#> 8           6       0       28 28  <NA>        4 5: 13, 4: 11, 1: 2, 3: 2  <NA>
#> 9           7       0       28 28  <NA>        4  2: 11, 4: 9, 3: 7, 5: 1  <NA>
#> 10          8       0       28 28  <NA>        4  5: 16, 4: 8, 3: 3, 2: 1  <NA>
#> 11          9       0       28 28  <NA>        4  3: 10, 4: 9, 2: 8, 5: 1  <NA>
#> 12         10       0       28 28  <NA>        5  4: 12, 3: 7, 2: 4, 5: 4  <NA>
#> 13         11       0       28 28  <NA>        4  4: 15, 3: 7, 5: 5, 2: 1  <NA>
#> 14         12       0       28 28  <NA>        5  4: 12, 1: 5, 3: 5, 2: 3  <NA>
#> 15         13       0       28 28  <NA>        5   2: 9, 4: 8, 3: 4, 5: 4  <NA>
#> 16         14       0       28 28  <NA>        5   4: 9, 5: 9, 3: 6, 1: 2  <NA>
#> 17         15       0       28 28  <NA>        5   3: 8, 2: 7, 4: 5, 5: 5  <NA>
#> 18         16       0       28 28  <NA>        4   2: 9, 4: 9, 1: 7, 3: 3  <NA>
#> 19         17       0       28 28  <NA>        5   4: 9, 5: 7, 2: 5, 3: 5  <NA>
#> 20         18       0       28 28  <NA>        4 4: 12, 5: 10, 3: 4, 2: 2  <NA>
#> 21         19       0       28 28  <NA>        4 5: 13, 4: 10, 3: 4, 1: 1  <NA>
#> 22         20       0       28 28  <NA>        5  4: 11, 5: 9, 3: 4, 2: 3  <NA>
#> 23         21       0       28 28  <NA>        5   4: 8, 5: 8, 3: 6, 2: 5  <NA>
#> 24         22       0       28 28  <NA>        4  5: 14, 4: 8, 3: 4, 2: 2  <NA>
#> 25       NULL       0       28 28  <NA>     <NA>                     <NA>  <NA>
#> 26       NULL       0       28 28  <NA>     <NA>                     <NA>  <NA>
#> 27       NULL       0       28 28  <NA>     <NA>                     <NA>  <NA>
#> 28       NULL       0       28 28  <NA>     <NA>                     <NA>  <NA>
#> 29       NULL       0       28 28  <NA>     <NA>                     <NA>  <NA>
#>        median        min        max mean   sd   p0  p25  p50  p75 p100     hist
#> 1        <NA>         64         64 <NA> <NA> <NA> <NA> <NA> <NA> <NA>     <NA>
#> 2  2016-07-08 2016-07-08 2016-11-02 <NA> <NA> <NA> <NA> <NA> <NA> <NA>     <NA>
#> 3  2016-07-08 2016-07-08 2016-11-02 <NA> <NA> <NA> <NA> <NA> <NA> <NA>     <NA>
#> 4  2016-07-08 2016-07-08 2016-11-02 <NA> <NA> <NA> <NA> <NA> <NA> <NA>     <NA>
#> 5        <NA>       <NA>       <NA>  NaN <NA> <NA> <NA> <NA> <NA> <NA>     <NA>
#> 6        <NA>       <NA>       <NA> <NA> <NA> <NA> <NA> <NA> <NA> <NA>     <NA>
#> 7        <NA>       <NA>       <NA> <NA> <NA> <NA> <NA> <NA> <NA> <NA>     <NA>
#> 8        <NA>       <NA>       <NA> <NA> <NA> <NA> <NA> <NA> <NA> <NA>     <NA>
#> 9        <NA>       <NA>       <NA> <NA> <NA> <NA> <NA> <NA> <NA> <NA>     <NA>
#> 10       <NA>       <NA>       <NA> <NA> <NA> <NA> <NA> <NA> <NA> <NA>     <NA>
#> 11       <NA>       <NA>       <NA> <NA> <NA> <NA> <NA> <NA> <NA> <NA>     <NA>
#> 12       <NA>       <NA>       <NA> <NA> <NA> <NA> <NA> <NA> <NA> <NA>     <NA>
#> 13       <NA>       <NA>       <NA> <NA> <NA> <NA> <NA> <NA> <NA> <NA>     <NA>
#> 14       <NA>       <NA>       <NA> <NA> <NA> <NA> <NA> <NA> <NA> <NA>     <NA>
#> 15       <NA>       <NA>       <NA> <NA> <NA> <NA> <NA> <NA> <NA> <NA>     <NA>
#> 16       <NA>       <NA>       <NA> <NA> <NA> <NA> <NA> <NA> <NA> <NA>     <NA>
#> 17       <NA>       <NA>       <NA> <NA> <NA> <NA> <NA> <NA> <NA> <NA>     <NA>
#> 18       <NA>       <NA>       <NA> <NA> <NA> <NA> <NA> <NA> <NA> <NA>     <NA>
#> 19       <NA>       <NA>       <NA> <NA> <NA> <NA> <NA> <NA> <NA> <NA>     <NA>
#> 20       <NA>       <NA>       <NA> <NA> <NA> <NA> <NA> <NA> <NA> <NA>     <NA>
#> 21       <NA>       <NA>       <NA> <NA> <NA> <NA> <NA> <NA> <NA> <NA>     <NA>
#> 22       <NA>       <NA>       <NA> <NA> <NA> <NA> <NA> <NA> <NA> <NA>     <NA>
#> 23       <NA>       <NA>       <NA> <NA> <NA> <NA> <NA> <NA> <NA> <NA>     <NA>
#> 24       <NA>       <NA>       <NA> <NA> <NA> <NA> <NA> <NA> <NA> <NA>     <NA>
#> 25       <NA>       <NA>       <NA> 3.12 0.93 1.5  2.25 3    4    4.75 ▂▇▂▇▃▁▅▅
#> 26       <NA>       <NA>       <NA> 4.26 0.56 3    3.94 4.25 4.75 5    ▃▂▂▁▇▃▇▂
#> 27       <NA>       <NA>       <NA> 3.65 0.79 2    3.12 3.75 4.25 5    ▂▃▁▃▃▇▂▂
#> 28       <NA>       <NA>       <NA> 3.85 1.01 1.5  3.25 4.25 4.56 5    ▂▁▂▃▁▃▇▇
#> 29       <NA>       <NA>       <NA> 2.89 0.93 1.33 2    2.83 3.42 4.33 ▅▅▂▅▃▅▁▇