Browse and search variable and value labels

Same as the codebook_browser(), but doesn't show data summaries and additional attributes. This yields a static table, so you can continue to edit code while viewing the labels, but you cannot switch the dataset via a dropdown menu.

Usage

label_browser_static(data = NULL, viewer = rstudioapi::viewer)

Arguments

data

data frame. if left empty, will use the text you currently select in RStudio as the label or the first data frame in your environment

viewer

where to show. defaults to viewer tab

Examples

label_browser_static(bfi)