Metadata as JSON-LD
Source: R/codebook.R

Echo a list of a metadata, generated using metadata_list() as JSON-LD in a script tag.

metadata_jsonld(results)

Arguments

results
a data frame, ideally with attributes set on variables

Examples

data("bfi")
metadata_jsonld(bfi)

#> 
#> 
#> 
#> <script type="application/ld+json">
#> {
#>   "name": "results",
#>   "keywords": ["session", "created", "modified", "ended", "expired", "BFIK_open_2", "BFIK_agree_4R", "BFIK_extra_2", "BFIK_agree_1R", "BFIK_open_1", "BFIK_neuro_2R", "BFIK_consc_3", "BFIK_consc_4", "BFIK_consc_2R", "BFIK_agree_3R", "BFIK_extra_3R", "BFIK_neuro_3", "BFIK_neuro_4", "BFIK_agree_2", "BFIK_consc_1", "BFIK_open_4", "BFIK_extra_4", "BFIK_extra_1R", "BFIK_open_3", "BFIK_agree", "BFIK_open", "BFIK_consc", "BFIK_extra", "BFIK_neuro"],
#>   "variableMeasured": [
#>     {
#>       "name": "session",
#>       "data_summary": [
#>         {
#>           "missing": "0",
#>           "complete": "28",
#>           "n": "28",
#>           "min": "64",
#>           "max": "64",
#>           "empty": "0",
#>           "n_unique": "28",
#>           "@type": "http://formr.org/codebook/SummaryStatistics"
#>         }
#>       ],
#>       "value": "text",
#>       "@context": {
#>         "@vocab": "http://pending.schema.org/"
#>       },
#>       "@type": "propertyValue"
#>     },
#>     {
#>       "name": "created",
#>       "description": "user first opened survey",
#>       "data_summary": [
#>         {
#>           "missing": "0",
#>           "complete": "28",
#>           "n": "28",
#>           "min": "2016-07-08",
#>           "max": "2016-11-02",
#>           "median": "2016-07-08",
#>           "n_unique": "28",
#>           "@type": "http://formr.org/codebook/SummaryStatistics"
#>         }
#>       ],
#>       "@context": {
#>         "@vocab": "http://pending.schema.org/"
#>       },
#>       "@type": "propertyValue"
#>     },
#>     {
#>       "name": "modified",
#>       "description": "user last edited survey",
#>       "data_summary": [
#>         {
#>           "missing": "0",
#>           "complete": "28",
#>           "n": "28",
#>           "min": "2016-07-08",
#>           "max": "2016-11-02",
#>           "median": "2016-07-08",
#>           "n_unique": "28",
#>           "@type": "http://formr.org/codebook/SummaryStatistics"
#>         }
#>       ],
#>       "@context": {
#>         "@vocab": "http://pending.schema.org/"
#>       },
#>       "@type": "propertyValue"
#>     },
#>     {
#>       "name": "ended",
#>       "description": "user finished survey",
#>       "data_summary": [
#>         {
#>           "missing": "0",
#>           "complete": "28",
#>           "n": "28",
#>           "min": "2016-07-08",
#>           "max": "2016-11-02",
#>           "median": "2016-07-08",
#>           "n_unique": "28",
#>           "@type": "http://formr.org/codebook/SummaryStatistics"
#>         }
#>       ],
#>       "@context": {
#>         "@vocab": "http://pending.schema.org/"
#>       },
#>       "@type": "propertyValue"
#>     },
#>     {
#>       "name": "expired",
#>       "data_summary": [
#>         {
#>           "missing": "28",
#>           "complete": "0",
#>           "n": "28",
#>           "mean": "NaN",
#>           "count": "28",
#>           "@type": "http://formr.org/codebook/SummaryStatistics"
#>         }
#>       ],
#>       "@context": {
#>         "@vocab": "http://pending.schema.org/"
#>       },
#>       "@type": "propertyValue"
#>     },
#>     {
#>       "item": {
#>         "choice_list": "BFIK_open_2",
#>         "type_options": 5,
#>         "name": "BFIK_open_2",
#>         "label": "__Ich bin tiefsinnig, denke gerne über Sachen nach.__",
#>         "label_parsed": "<strong>Ich bin tiefsinnig, denke gerne über Sachen nach.<\/strong>",
#>         "optional": 0,
#>         "class": {},
#>         "showif": {},
#>         "value": {},
#>         "block_order": {},
#>         "item_order": 4,
#>         "choices": {
#>           "1": "1: Trifft überhaupt nicht zu",
#>           "2": "2",
#>           "3": "3",
#>           "4": "4",
#>           "5": "5: Trifft voll und ganz zu",
#>           "@type": "http://formr.org/codebook/ItemChoices"
#>         },
#>         "item_type": "rating_button",
#>         "@type": "http://formr.org/codebook/Item"
#>       },
#>       "name": "BFIK_open_2",
#>       "description": "__Ich bin tiefsinnig, denke gerne über Sachen nach.__",
#>       "value": {
#>         "1: Trifft überhaupt nicht zu": 1,
#>         "2": 2,
#>         "3": 3,
#>         "4": 4,
#>         "5: Trifft voll und ganz zu": 5,
#>         "Item was never rendered for this user.": "NA"
#>       },
#>       "maxValue": 5,
#>       "minValue": 1,
#>       "measurementTechnique": "self-report",
#>       "data_summary": [
#>         {
#>           "missing": "0",
#>           "complete": "28",
#>           "n": "28",
#>           "n_unique": "4",
#>           "top_counts": "4: 15, 5: 10, 3: 2, 2: 1",
#>           "ordered": "FALSE",
#>           "@type": "http://formr.org/codebook/SummaryStatistics"
#>         }
#>       ],
#>       "@context": {
#>         "@vocab": "http://pending.schema.org/"
#>       },
#>       "@type": "propertyValue"
#>     },
#>     {
#>       "item": {
#>         "choice_list": "BFIK_agree_4R",
#>         "type_options": 5,
#>         "name": "BFIK_agree_4R",
#>         "label": "__Ich kann mich schroff und abweisend anderen gegenüber verhalten.__",
#>         "label_parsed": "<strong>Ich kann mich schroff und abweisend anderen gegenüber verhalten.<\/strong>",
#>         "optional": 0,
#>         "class": {},
#>         "showif": {},
#>         "value": {},
#>         "block_order": {},
#>         "item_order": 5,
#>         "choices": {
#>           "1": "1: Trifft überhaupt nicht zu",
#>           "2": "2",
#>           "3": "3",
#>           "4": "4",
#>           "5": "5: Trifft voll und ganz zu",
#>           "@type": "http://formr.org/codebook/ItemChoices"
#>         },
#>         "item_type": "rating_button",
#>         "@type": "http://formr.org/codebook/Item"
#>       },
#>       "name": "BFIK_agree_4R",
#>       "description": "__Ich kann mich schroff und abweisend anderen gegenüber verhalten.__",
#>       "value": {
#>         "1: Trifft überhaupt nicht zu": 5,
#>         "2": 4,
#>         "3": 3,
#>         "4": 2,
#>         "5: Trifft voll und ganz zu": 1,
#>         "Item was never rendered for this user.": "NA"
#>       },
#>       "maxValue": 5,
#>       "minValue": 1,
#>       "measurementTechnique": "self-report",
#>       "data_summary": [
#>         {
#>           "missing": "0",
#>           "complete": "28",
#>           "n": "28",
#>           "n_unique": "5",
#>           "top_counts": "2: 11, 4: 7, 3: 5, 5: 3",
#>           "ordered": "FALSE",
#>           "@type": "http://formr.org/codebook/SummaryStatistics"
#>         }
#>       ],
#>       "@context": {
#>         "@vocab": "http://pending.schema.org/"
#>       },
#>       "@type": "propertyValue"
#>     },
#>     {
#>       "item": {
#>         "choice_list": "BFIK_extra_2",
#>         "type_options": 5,
#>         "name": "BFIK_extra_2",
#>         "label": "__Ich bin begeisterungsfähig und kann andere leicht mitreißen.__",
#>         "label_parsed": "<strong>Ich bin begeisterungsfähig und kann andere leicht mitreißen.<\/strong>",
#>         "optional": 0,
#>         "class": {},
#>         "showif": {},
#>         "value": {},
#>         "block_order": {},
#>         "item_order": 6,
#>         "choices": {
#>           "1": "1: Trifft überhaupt nicht zu",
#>           "2": "2",
#>           "3": "3",
#>           "4": "4",
#>           "5": "5: Trifft voll und ganz zu",
#>           "@type": "http://formr.org/codebook/ItemChoices"
#>         },
#>         "item_type": "rating_button",
#>         "@type": "http://formr.org/codebook/Item"
#>       },
#>       "name": "BFIK_extra_2",
#>       "description": "__Ich bin begeisterungsfähig und kann andere leicht mitreißen.__",
#>       "value": {
#>         "1: Trifft überhaupt nicht zu": 1,
#>         "2": 2,
#>         "3": 3,
#>         "4": 4,
#>         "5: Trifft voll und ganz zu": 5,
#>         "Item was never rendered for this user.": "NA"
#>       },
#>       "maxValue": 5,
#>       "minValue": 1,
#>       "measurementTechnique": "self-report",
#>       "data_summary": [
#>         {
#>           "missing": "0",
#>           "complete": "28",
#>           "n": "28",
#>           "n_unique": "4",
#>           "top_counts": "5: 13, 4: 11, 1: 2, 3: 2",
#>           "ordered": "FALSE",
#>           "@type": "http://formr.org/codebook/SummaryStatistics"
#>         }
#>       ],
#>       "@context": {
#>         "@vocab": "http://pending.schema.org/"
#>       },
#>       "@type": "propertyValue"
#>     },
#>     {
#>       "item": {
#>         "choice_list": "BFIK_agree_1R",
#>         "type_options": 5,
#>         "name": "BFIK_agree_1R",
#>         "label": "__Ich neige dazu, andere zu kritisieren.__",
#>         "label_parsed": "<strong>Ich neige dazu, andere zu kritisieren.<\/strong>",
#>         "optional": 0,
#>         "class": {},
#>         "showif": {},
#>         "value": {},
#>         "block_order": {},
#>         "item_order": 7,
#>         "choices": {
#>           "1": "1: Trifft überhaupt nicht zu",
#>           "2": "2",
#>           "3": "3",
#>           "4": "4",
#>           "5": "5: Trifft voll und ganz zu",
#>           "@type": "http://formr.org/codebook/ItemChoices"
#>         },
#>         "item_type": "rating_button",
#>         "@type": "http://formr.org/codebook/Item"
#>       },
#>       "name": "BFIK_agree_1R",
#>       "description": "__Ich neige dazu, andere zu kritisieren.__",
#>       "value": {
#>         "1: Trifft überhaupt nicht zu": 5,
#>         "2": 4,
#>         "3": 3,
#>         "4": 2,
#>         "5: Trifft voll und ganz zu": 1,
#>         "Item was never rendered for this user.": "NA"
#>       },
#>       "maxValue": 5,
#>       "minValue": 1,
#>       "measurementTechnique": "self-report",
#>       "data_summary": [
#>         {
#>           "missing": "0",
#>           "complete": "28",
#>           "n": "28",
#>           "n_unique": "4",
#>           "top_counts": "2: 11, 4: 9, 3: 7, 5: 1",
#>           "ordered": "FALSE",
#>           "@type": "http://formr.org/codebook/SummaryStatistics"
#>         }
#>       ],
#>       "@context": {
#>         "@vocab": "http://pending.schema.org/"
#>       },
#>       "@type": "propertyValue"
#>     },
#>     {
#>       "item": {
#>         "choice_list": "BFIK_open_1",
#>         "type_options": 5,
#>         "name": "BFIK_open_1",
#>         "label": "__Ich bin vielseitig interessiert.__",
#>         "label_parsed": "<strong>Ich bin vielseitig interessiert.<\/strong>",
#>         "optional": 0,
#>         "class": {},
#>         "showif": {},
#>         "value": {},
#>         "block_order": {},
#>         "item_order": 8,
#>         "choices": {
#>           "1": "1: Trifft überhaupt nicht zu",
#>           "2": "2",
#>           "3": "3",
#>           "4": "4",
#>           "5": "5: Trifft voll und ganz zu",
#>           "@type": "http://formr.org/codebook/ItemChoices"
#>         },
#>         "item_type": "rating_button",
#>         "@type": "http://formr.org/codebook/Item"
#>       },
#>       "name": "BFIK_open_1",
#>       "description": "__Ich bin vielseitig interessiert.__",
#>       "value": {
#>         "1: Trifft überhaupt nicht zu": 1,
#>         "2": 2,
#>         "3": 3,
#>         "4": 4,
#>         "5: Trifft voll und ganz zu": 5,
#>         "Item was never rendered for this user.": "NA"
#>       },
#>       "maxValue": 5,
#>       "minValue": 1,
#>       "measurementTechnique": "self-report",
#>       "data_summary": [
#>         {
#>           "missing": "0",
#>           "complete": "28",
#>           "n": "28",
#>           "n_unique": "4",
#>           "top_counts": "5: 16, 4: 8, 3: 3, 2: 1",
#>           "ordered": "FALSE",
#>           "@type": "http://formr.org/codebook/SummaryStatistics"
#>         }
#>       ],
#>       "@context": {
#>         "@vocab": "http://pending.schema.org/"
#>       },
#>       "@type": "propertyValue"
#>     },
#>     {
#>       "item": {
#>         "choice_list": "BFIK_neuro_2R",
#>         "type_options": 5,
#>         "name": "BFIK_neuro_2R",
#>         "label": "__Ich bin entspannt, lasse mich durch Stress nicht aus der Ruhe bringen.__",
#>         "label_parsed": "<strong>Ich bin entspannt, lasse mich durch Stress nicht aus der Ruhe bringen.<\/strong>",
#>         "optional": 0,
#>         "class": {},
#>         "showif": {},
#>         "value": {},
#>         "block_order": {},
#>         "item_order": 9,
#>         "choices": {
#>           "1": "1: Trifft überhaupt nicht zu",
#>           "2": "2",
#>           "3": "3",
#>           "4": "4",
#>           "5": "5: Trifft voll und ganz zu",
#>           "@type": "http://formr.org/codebook/ItemChoices"
#>         },
#>         "item_type": "rating_button",
#>         "@type": "http://formr.org/codebook/Item"
#>       },
#>       "name": "BFIK_neuro_2R",
#>       "description": "__Ich bin entspannt, lasse mich durch Stress nicht aus der Ruhe bringen.__",
#>       "value": {
#>         "1: Trifft überhaupt nicht zu": 5,
#>         "2": 4,
#>         "3": 3,
#>         "4": 2,
#>         "5: Trifft voll und ganz zu": 1,
#>         "Item was never rendered for this user.": "NA"
#>       },
#>       "maxValue": 5,
#>       "minValue": 1,
#>       "measurementTechnique": "self-report",
#>       "data_summary": [
#>         {
#>           "missing": "0",
#>           "complete": "28",
#>           "n": "28",
#>           "n_unique": "4",
#>           "top_counts": "3: 10, 4: 9, 2: 8, 5: 1",
#>           "ordered": "FALSE",
#>           "@type": "http://formr.org/codebook/SummaryStatistics"
#>         }
#>       ],
#>       "@context": {
#>         "@vocab": "http://pending.schema.org/"
#>       },
#>       "@type": "propertyValue"
#>     },
#>     {
#>       "item": {
#>         "choice_list": "BFIK_consc_3",
#>         "type_options": 5,
#>         "name": "BFIK_consc_3",
#>         "label": "__Ich bin tüchtig und arbeite flott.__",
#>         "label_parsed": "<strong>Ich bin tüchtig und arbeite flott.<\/strong>",
#>         "optional": 0,
#>         "class": {},
#>         "showif": {},
#>         "value": {},
#>         "block_order": {},
#>         "item_order": 10,
#>         "choices": {
#>           "1": "1: Trifft überhaupt nicht zu",
#>           "2": "2",
#>           "3": "3",
#>           "4": "4",
#>           "5": "5: Trifft voll und ganz zu",
#>           "@type": "http://formr.org/codebook/ItemChoices"
#>         },
#>         "item_type": "rating_button",
#>         "@type": "http://formr.org/codebook/Item"
#>       },
#>       "name": "BFIK_consc_3",
#>       "description": "__Ich bin tüchtig und arbeite flott.__",
#>       "value": {
#>         "1: Trifft überhaupt nicht zu": 1,
#>         "2": 2,
#>         "3": 3,
#>         "4": 4,
#>         "5: Trifft voll und ganz zu": 5,
#>         "Item was never rendered for this user.": "NA"
#>       },
#>       "maxValue": 5,
#>       "minValue": 1,
#>       "measurementTechnique": "self-report",
#>       "data_summary": [
#>         {
#>           "missing": "0",
#>           "complete": "28",
#>           "n": "28",
#>           "n_unique": "5",
#>           "top_counts": "4: 12, 3: 7, 2: 4, 5: 4",
#>           "ordered": "FALSE",
#>           "@type": "http://formr.org/codebook/SummaryStatistics"
#>         }
#>       ],
#>       "@context": {
#>         "@vocab": "http://pending.schema.org/"
#>       },
#>       "@type": "propertyValue"
#>     },
#>     {
#>       "item": {
#>         "choice_list": "BFIK_consc_4",
#>         "type_options": 5,
#>         "name": "BFIK_consc_4",
#>         "label": "__Ich mache Pläne und führe sie auch durch.__",
#>         "label_parsed": "<strong>Ich mache Pläne und führe sie auch durch.<\/strong>",
#>         "optional": 0,
#>         "class": {},
#>         "showif": {},
#>         "value": {},
#>         "block_order": {},
#>         "item_order": 11,
#>         "choices": {
#>           "1": "1: Trifft überhaupt nicht zu",
#>           "2": "2",
#>           "3": "3",
#>           "4": "4",
#>           "5": "5: Trifft voll und ganz zu",
#>           "@type": "http://formr.org/codebook/ItemChoices"
#>         },
#>         "item_type": "rating_button",
#>         "@type": "http://formr.org/codebook/Item"
#>       },
#>       "name": "BFIK_consc_4",
#>       "description": "__Ich mache Pläne und führe sie auch durch.__",
#>       "value": {
#>         "1: Trifft überhaupt nicht zu": 1,
#>         "2": 2,
#>         "3": 3,
#>         "4": 4,
#>         "5: Trifft voll und ganz zu": 5,
#>         "Item was never rendered for this user.": "NA"
#>       },
#>       "maxValue": 5,
#>       "minValue": 1,
#>       "measurementTechnique": "self-report",
#>       "data_summary": [
#>         {
#>           "missing": "0",
#>           "complete": "28",
#>           "n": "28",
#>           "n_unique": "4",
#>           "top_counts": "4: 15, 3: 7, 5: 5, 2: 1",
#>           "ordered": "FALSE",
#>           "@type": "http://formr.org/codebook/SummaryStatistics"
#>         }
#>       ],
#>       "@context": {
#>         "@vocab": "http://pending.schema.org/"
#>       },
#>       "@type": "propertyValue"
#>     },
#>     {
#>       "item": {
#>         "choice_list": "BFIK_consc_2R",
#>         "type_options": 5,
#>         "name": "BFIK_consc_2R",
#>         "label": "__Ich bin bequem, neige zur Faulheit.__",
#>         "label_parsed": "<strong>Ich bin bequem, neige zur Faulheit.<\/strong>",
#>         "optional": 0,
#>         "class": {},
#>         "showif": {},
#>         "value": {},
#>         "block_order": {},
#>         "item_order": 12,
#>         "choices": {
#>           "1": "1: Trifft überhaupt nicht zu",
#>           "2": "2",
#>           "3": "3",
#>           "4": "4",
#>           "5": "5: Trifft voll und ganz zu",
#>           "@type": "http://formr.org/codebook/ItemChoices"
#>         },
#>         "item_type": "rating_button",
#>         "@type": "http://formr.org/codebook/Item"
#>       },
#>       "name": "BFIK_consc_2R",
#>       "description": "__Ich bin bequem, neige zur Faulheit.__",
#>       "value": {
#>         "1: Trifft überhaupt nicht zu": 5,
#>         "2": 4,
#>         "3": 3,
#>         "4": 2,
#>         "5: Trifft voll und ganz zu": 1,
#>         "Item was never rendered for this user.": "NA"
#>       },
#>       "maxValue": 5,
#>       "minValue": 1,
#>       "measurementTechnique": "self-report",
#>       "data_summary": [
#>         {
#>           "missing": "0",
#>           "complete": "28",
#>           "n": "28",
#>           "n_unique": "5",
#>           "top_counts": "4: 12, 1: 5, 3: 5, 2: 3",
#>           "ordered": "FALSE",
#>           "@type": "http://formr.org/codebook/SummaryStatistics"
#>         }
#>       ],
#>       "@context": {
#>         "@vocab": "http://pending.schema.org/"
#>       },
#>       "@type": "propertyValue"
#>     },
#>     {
#>       "item": {
#>         "choice_list": "BFIK_agree_3R",
#>         "type_options": 5,
#>         "name": "BFIK_agree_3R",
#>         "label": "__Ich kann mich kalt und distanziert verhalten.__",
#>         "label_parsed": "<strong>Ich kann mich kalt und distanziert verhalten.<\/strong>",
#>         "optional": 0,
#>         "class": {},
#>         "showif": {},
#>         "value": {},
#>         "block_order": {},
#>         "item_order": 13,
#>         "choices": {
#>           "1": "1: Trifft überhaupt nicht zu",
#>           "2": "2",
#>           "3": "3",
#>           "4": "4",
#>           "5": "5: Trifft voll und ganz zu",
#>           "@type": "http://formr.org/codebook/ItemChoices"
#>         },
#>         "item_type": "rating_button",
#>         "@type": "http://formr.org/codebook/Item"
#>       },
#>       "name": "BFIK_agree_3R",
#>       "description": "__Ich kann mich kalt und distanziert verhalten.__",
#>       "value": {
#>         "1: Trifft überhaupt nicht zu": 5,
#>         "2": 4,
#>         "3": 3,
#>         "4": 2,
#>         "5: Trifft voll und ganz zu": 1,
#>         "Item was never rendered for this user.": "NA"
#>       },
#>       "maxValue": 5,
#>       "minValue": 1,
#>       "measurementTechnique": "self-report",
#>       "data_summary": [
#>         {
#>           "missing": "0",
#>           "complete": "28",
#>           "n": "28",
#>           "n_unique": "5",
#>           "top_counts": "2: 9, 4: 8, 3: 4, 5: 4",
#>           "ordered": "FALSE",
#>           "@type": "http://formr.org/codebook/SummaryStatistics"
#>         }
#>       ],
#>       "@context": {
#>         "@vocab": "http://pending.schema.org/"
#>       },
#>       "@type": "propertyValue"
#>     },
#>     {
#>       "item": {
#>         "choice_list": "BFIK_extra_3R",
#>         "type_options": 5,
#>         "name": "BFIK_extra_3R",
#>         "label": "__Ich bin eher der \"stille Typ\", wortkarg.__",
#>         "label_parsed": "<strong>Ich bin eher der &quot;stille Typ&quot;, wortkarg.<\/strong>",
#>         "optional": 0,
#>         "class": {},
#>         "showif": {},
#>         "value": {},
#>         "block_order": {},
#>         "item_order": 14,
#>         "choices": {
#>           "1": "1: Trifft überhaupt nicht zu",
#>           "2": "2",
#>           "3": "3",
#>           "4": "4",
#>           "5": "5: Trifft voll und ganz zu",
#>           "@type": "http://formr.org/codebook/ItemChoices"
#>         },
#>         "item_type": "rating_button",
#>         "@type": "http://formr.org/codebook/Item"
#>       },
#>       "name": "BFIK_extra_3R",
#>       "description": "__Ich bin eher der \"stille Typ\", wortkarg.__",
#>       "value": {
#>         "1: Trifft überhaupt nicht zu": 5,
#>         "2": 4,
#>         "3": 3,
#>         "4": 2,
#>         "5: Trifft voll und ganz zu": 1,
#>         "Item was never rendered for this user.": "NA"
#>       },
#>       "maxValue": 5,
#>       "minValue": 1,
#>       "measurementTechnique": "self-report",
#>       "data_summary": [
#>         {
#>           "missing": "0",
#>           "complete": "28",
#>           "n": "28",
#>           "n_unique": "5",
#>           "top_counts": "4: 9, 5: 9, 3: 6, 1: 2",
#>           "ordered": "FALSE",
#>           "@type": "http://formr.org/codebook/SummaryStatistics"
#>         }
#>       ],
#>       "@context": {
#>         "@vocab": "http://pending.schema.org/"
#>       },
#>       "@type": "propertyValue"
#>     },
#>     {
#>       "item": {
#>         "choice_list": "BFIK_neuro_3",
#>         "type_options": 5,
#>         "name": "BFIK_neuro_3",
#>         "label": "__Ich mache mir viele Sorgen.__",
#>         "label_parsed": "<strong>Ich mache mir viele Sorgen.<\/strong>",
#>         "optional": 0,
#>         "class": {},
#>         "showif": {},
#>         "value": {},
#>         "block_order": {},
#>         "item_order": 15,
#>         "choices": {
#>           "1": "1: Trifft überhaupt nicht zu",
#>           "2": "2",
#>           "3": "3",
#>           "4": "4",
#>           "5": "5: Trifft voll und ganz zu",
#>           "@type": "http://formr.org/codebook/ItemChoices"
#>         },
#>         "item_type": "rating_button",
#>         "@type": "http://formr.org/codebook/Item"
#>       },
#>       "name": "BFIK_neuro_3",
#>       "description": "__Ich mache mir viele Sorgen.__",
#>       "value": {
#>         "1: Trifft überhaupt nicht zu": 1,
#>         "2": 2,
#>         "3": 3,
#>         "4": 4,
#>         "5: Trifft voll und ganz zu": 5,
#>         "Item was never rendered for this user.": "NA"
#>       },
#>       "maxValue": 5,
#>       "minValue": 1,
#>       "measurementTechnique": "self-report",
#>       "data_summary": [
#>         {
#>           "missing": "0",
#>           "complete": "28",
#>           "n": "28",
#>           "n_unique": "5",
#>           "top_counts": "3: 8, 2: 7, 4: 5, 5: 5",
#>           "ordered": "FALSE",
#>           "@type": "http://formr.org/codebook/SummaryStatistics"
#>         }
#>       ],
#>       "@context": {
#>         "@vocab": "http://pending.schema.org/"
#>       },
#>       "@type": "propertyValue"
#>     },
#>     {
#>       "item": {
#>         "choice_list": "BFIK_neuro_4",
#>         "type_options": 5,
#>         "name": "BFIK_neuro_4",
#>         "label": "__Ich werde leicht nervös und unsicher.__",
#>         "label_parsed": "<strong>Ich werde leicht nervös und unsicher.<\/strong>",
#>         "optional": 0,
#>         "class": {},
#>         "showif": {},
#>         "value": {},
#>         "block_order": {},
#>         "item_order": 16,
#>         "choices": {
#>           "1": "1: Trifft überhaupt nicht zu",
#>           "2": "2",
#>           "3": "3",
#>           "4": "4",
#>           "5": "5: Trifft voll und ganz zu",
#>           "@type": "http://formr.org/codebook/ItemChoices"
#>         },
#>         "item_type": "rating_button",
#>         "@type": "http://formr.org/codebook/Item"
#>       },
#>       "name": "BFIK_neuro_4",
#>       "description": "__Ich werde leicht nervös und unsicher.__",
#>       "value": {
#>         "1: Trifft überhaupt nicht zu": 1,
#>         "2": 2,
#>         "3": 3,
#>         "4": 4,
#>         "5: Trifft voll und ganz zu": 5,
#>         "Item was never rendered for this user.": "NA"
#>       },
#>       "maxValue": 5,
#>       "minValue": 1,
#>       "measurementTechnique": "self-report",
#>       "data_summary": [
#>         {
#>           "missing": "0",
#>           "complete": "28",
#>           "n": "28",
#>           "n_unique": "4",
#>           "top_counts": "2: 9, 4: 9, 1: 7, 3: 3",
#>           "ordered": "FALSE",
#>           "@type": "http://formr.org/codebook/SummaryStatistics"
#>         }
#>       ],
#>       "@context": {
#>         "@vocab": "http://pending.schema.org/"
#>       },
#>       "@type": "propertyValue"
#>     },
#>     {
#>       "item": {
#>         "choice_list": "BFIK_agree_2",
#>         "type_options": 5,
#>         "name": "BFIK_agree_2",
#>         "label": "__Ich schenke anderen leicht Vertrauen, glaube an das Gute im Menschen.__",
#>         "label_parsed": "<strong>Ich schenke anderen leicht Vertrauen, glaube an das Gute im Menschen.<\/strong>",
#>         "optional": 0,
#>         "class": {},
#>         "showif": {},
#>         "value": {},
#>         "block_order": {},
#>         "item_order": 17,
#>         "choices": {
#>           "1": "1: Trifft überhaupt nicht zu",
#>           "2": "2",
#>           "3": "3",
#>           "4": "4",
#>           "5": "5: Trifft voll und ganz zu",
#>           "@type": "http://formr.org/codebook/ItemChoices"
#>         },
#>         "item_type": "rating_button",
#>         "@type": "http://formr.org/codebook/Item"
#>       },
#>       "name": "BFIK_agree_2",
#>       "description": "__Ich schenke anderen leicht Vertrauen, glaube an das Gute im Menschen.__",
#>       "value": {
#>         "1: Trifft überhaupt nicht zu": 1,
#>         "2": 2,
#>         "3": 3,
#>         "4": 4,
#>         "5: Trifft voll und ganz zu": 5,
#>         "Item was never rendered for this user.": "NA"
#>       },
#>       "maxValue": 5,
#>       "minValue": 1,
#>       "measurementTechnique": "self-report",
#>       "data_summary": [
#>         {
#>           "missing": "0",
#>           "complete": "28",
#>           "n": "28",
#>           "n_unique": "5",
#>           "top_counts": "4: 9, 5: 7, 2: 5, 3: 5",
#>           "ordered": "FALSE",
#>           "@type": "http://formr.org/codebook/SummaryStatistics"
#>         }
#>       ],
#>       "@context": {
#>         "@vocab": "http://pending.schema.org/"
#>       },
#>       "@type": "propertyValue"
#>     },
#>     {
#>       "item": {
#>         "choice_list": "BFIK_consc_1",
#>         "type_options": 5,
#>         "name": "BFIK_consc_1",
#>         "label": "__Ich erledige Aufgaben gründlich.__",
#>         "label_parsed": "<strong>Ich erledige Aufgaben gründlich.<\/strong>",
#>         "optional": 0,
#>         "class": {},
#>         "showif": {},
#>         "value": {},
#>         "block_order": {},
#>         "item_order": 18,
#>         "choices": {
#>           "1": "1: Trifft überhaupt nicht zu",
#>           "2": "2",
#>           "3": "3",
#>           "4": "4",
#>           "5": "5: Trifft voll und ganz zu",
#>           "@type": "http://formr.org/codebook/ItemChoices"
#>         },
#>         "item_type": "rating_button",
#>         "@type": "http://formr.org/codebook/Item"
#>       },
#>       "name": "BFIK_consc_1",
#>       "description": "__Ich erledige Aufgaben gründlich.__",
#>       "value": {
#>         "1: Trifft überhaupt nicht zu": 1,
#>         "2": 2,
#>         "3": 3,
#>         "4": 4,
#>         "5: Trifft voll und ganz zu": 5,
#>         "Item was never rendered for this user.": "NA"
#>       },
#>       "maxValue": 5,
#>       "minValue": 1,
#>       "measurementTechnique": "self-report",
#>       "data_summary": [
#>         {
#>           "missing": "0",
#>           "complete": "28",
#>           "n": "28",
#>           "n_unique": "4",
#>           "top_counts": "4: 12, 5: 10, 3: 4, 2: 2",
#>           "ordered": "FALSE",
#>           "@type": "http://formr.org/codebook/SummaryStatistics"
#>         }
#>       ],
#>       "@context": {
#>         "@vocab": "http://pending.schema.org/"
#>       },
#>       "@type": "propertyValue"
#>     },
#>     {
#>       "item": {
#>         "choice_list": "BFIK_open_4",
#>         "type_options": 5,
#>         "name": "BFIK_open_4",
#>         "label": "__Ich schätze künstlerische und ästhetische Eindrücke.__",
#>         "label_parsed": "<strong>Ich schätze künstlerische und ästhetische Eindrücke.<\/strong>",
#>         "optional": 0,
#>         "class": {},
#>         "showif": {},
#>         "value": {},
#>         "block_order": {},
#>         "item_order": 19,
#>         "choices": {
#>           "1": "1: Trifft überhaupt nicht zu",
#>           "2": "2",
#>           "3": "3",
#>           "4": "4",
#>           "5": "5: Trifft voll und ganz zu",
#>           "@type": "http://formr.org/codebook/ItemChoices"
#>         },
#>         "item_type": "rating_button",
#>         "@type": "http://formr.org/codebook/Item"
#>       },
#>       "name": "BFIK_open_4",
#>       "description": "__Ich schätze künstlerische und ästhetische Eindrücke.__",
#>       "value": {
#>         "1: Trifft überhaupt nicht zu": 1,
#>         "2": 2,
#>         "3": 3,
#>         "4": 4,
#>         "5: Trifft voll und ganz zu": 5,
#>         "Item was never rendered for this user.": "NA"
#>       },
#>       "maxValue": 5,
#>       "minValue": 1,
#>       "measurementTechnique": "self-report",
#>       "data_summary": [
#>         {
#>           "missing": "0",
#>           "complete": "28",
#>           "n": "28",
#>           "n_unique": "4",
#>           "top_counts": "5: 13, 4: 10, 3: 4, 1: 1",
#>           "ordered": "FALSE",
#>           "@type": "http://formr.org/codebook/SummaryStatistics"
#>         }
#>       ],
#>       "@context": {
#>         "@vocab": "http://pending.schema.org/"
#>       },
#>       "@type": "propertyValue"
#>     },
#>     {
#>       "item": {
#>         "choice_list": "BFIK_extra_4",
#>         "type_options": 5,
#>         "name": "BFIK_extra_4",
#>         "label": "__Ich gehe aus mir heraus, bin gesellig.__",
#>         "label_parsed": "<strong>Ich gehe aus mir heraus, bin gesellig.<\/strong>",
#>         "optional": 0,
#>         "class": {},
#>         "showif": {},
#>         "value": {},
#>         "block_order": {},
#>         "item_order": 20,
#>         "choices": {
#>           "1": "1: Trifft überhaupt nicht zu",
#>           "2": "2",
#>           "3": "3",
#>           "4": "4",
#>           "5": "5: Trifft voll und ganz zu",
#>           "@type": "http://formr.org/codebook/ItemChoices"
#>         },
#>         "item_type": "rating_button",
#>         "@type": "http://formr.org/codebook/Item"
#>       },
#>       "name": "BFIK_extra_4",
#>       "description": "__Ich gehe aus mir heraus, bin gesellig.__",
#>       "value": {
#>         "1: Trifft überhaupt nicht zu": 1,
#>         "2": 2,
#>         "3": 3,
#>         "4": 4,
#>         "5: Trifft voll und ganz zu": 5,
#>         "Item was never rendered for this user.": "NA"
#>       },
#>       "maxValue": 5,
#>       "minValue": 1,
#>       "measurementTechnique": "self-report",
#>       "data_summary": [
#>         {
#>           "missing": "0",
#>           "complete": "28",
#>           "n": "28",
#>           "n_unique": "5",
#>           "top_counts": "4: 11, 5: 9, 3: 4, 2: 3",
#>           "ordered": "FALSE",
#>           "@type": "http://formr.org/codebook/SummaryStatistics"
#>         }
#>       ],
#>       "@context": {
#>         "@vocab": "http://pending.schema.org/"
#>       },
#>       "@type": "propertyValue"
#>     },
#>     {
#>       "item": {
#>         "choice_list": "BFIK_extra_1R",
#>         "type_options": 5,
#>         "name": "BFIK_extra_1R",
#>         "label": "__Ich bin eher zurückhaltend, reserviert.__",
#>         "label_parsed": "<strong>Ich bin eher zurückhaltend, reserviert.<\/strong>",
#>         "optional": 0,
#>         "class": {},
#>         "showif": {},
#>         "value": {},
#>         "block_order": {},
#>         "item_order": 21,
#>         "choices": {
#>           "1": "1: Trifft überhaupt nicht zu",
#>           "2": "2",
#>           "3": "3",
#>           "4": "4",
#>           "5": "5: Trifft voll und ganz zu",
#>           "@type": "http://formr.org/codebook/ItemChoices"
#>         },
#>         "item_type": "rating_button",
#>         "@type": "http://formr.org/codebook/Item"
#>       },
#>       "name": "BFIK_extra_1R",
#>       "description": "__Ich bin eher zurückhaltend, reserviert.__",
#>       "value": {
#>         "1: Trifft überhaupt nicht zu": 5,
#>         "2": 4,
#>         "3": 3,
#>         "4": 2,
#>         "5: Trifft voll und ganz zu": 1,
#>         "Item was never rendered for this user.": "NA"
#>       },
#>       "maxValue": 5,
#>       "minValue": 1,
#>       "measurementTechnique": "self-report",
#>       "data_summary": [
#>         {
#>           "missing": "0",
#>           "complete": "28",
#>           "n": "28",
#>           "n_unique": "5",
#>           "top_counts": "4: 8, 5: 8, 3: 6, 2: 5",
#>           "ordered": "FALSE",
#>           "@type": "http://formr.org/codebook/SummaryStatistics"
#>         }
#>       ],
#>       "@context": {
#>         "@vocab": "http://pending.schema.org/"
#>       },
#>       "@type": "propertyValue"
#>     },
#>     {
#>       "item": {
#>         "choice_list": "BFIK_open_3",
#>         "type_options": 5,
#>         "name": "BFIK_open_3",
#>         "label": "__Ich habe eine aktive Vorstellungskraft, bin phantasievoll.__",
#>         "label_parsed": "<strong>Ich habe eine aktive Vorstellungskraft, bin phantasievoll.<\/strong>",
#>         "optional": 0,
#>         "class": {},
#>         "showif": {},
#>         "value": {},
#>         "block_order": {},
#>         "item_order": 22,
#>         "choices": {
#>           "1": "1: Trifft überhaupt nicht zu",
#>           "2": "2",
#>           "3": "3",
#>           "4": "4",
#>           "5": "5: Trifft voll und ganz zu",
#>           "@type": "http://formr.org/codebook/ItemChoices"
#>         },
#>         "item_type": "rating_button",
#>         "@type": "http://formr.org/codebook/Item"
#>       },
#>       "name": "BFIK_open_3",
#>       "description": "__Ich habe eine aktive Vorstellungskraft, bin phantasievoll.__",
#>       "value": {
#>         "1: Trifft überhaupt nicht zu": 1,
#>         "2": 2,
#>         "3": 3,
#>         "4": 4,
#>         "5: Trifft voll und ganz zu": 5,
#>         "Item was never rendered for this user.": "NA"
#>       },
#>       "maxValue": 5,
#>       "minValue": 1,
#>       "measurementTechnique": "self-report",
#>       "data_summary": [
#>         {
#>           "missing": "0",
#>           "complete": "28",
#>           "n": "28",
#>           "n_unique": "4",
#>           "top_counts": "5: 14, 4: 8, 3: 4, 2: 2",
#>           "ordered": "FALSE",
#>           "@type": "http://formr.org/codebook/SummaryStatistics"
#>         }
#>       ],
#>       "@context": {
#>         "@vocab": "http://pending.schema.org/"
#>       },
#>       "@type": "propertyValue"
#>     },
#>     {
#>       "scale_item_names": ["BFIK_agree_4R", "BFIK_agree_1R", "BFIK_agree_3R", "BFIK_agree_2"],
#>       "name": "BFIK_agree",
#>       "description": "4 BFIK_agree items averaged with aggregation_function",
#>       "data_summary": [
#>         {
#>           "missing": "0",
#>           "complete": "28",
#>           "n": "28",
#>           "mean": "3.12",
#>           "sd": "0.93",
#>           "p0": "1.5",
#>           "p25": "2.25",
#>           "p50": "3",
#>           "p75": "4",
#>           "p100": "4.75",
#>           "hist": "▂▇▂▇▃▁▅▅",
#>           "@type": "http://formr.org/codebook/SummaryStatistics"
#>         }
#>       ],
#>       "value": "Number",
#>       "@context": {
#>         "@vocab": "http://pending.schema.org/"
#>       },
#>       "@type": "propertyValue"
#>     },
#>     {
#>       "scale_item_names": ["BFIK_open_2", "BFIK_open_1", "BFIK_open_4", "BFIK_open_3"],
#>       "name": "BFIK_open",
#>       "description": "4 BFIK_open items averaged with aggregation_function",
#>       "data_summary": [
#>         {
#>           "missing": "0",
#>           "complete": "28",
#>           "n": "28",
#>           "mean": "4.26",
#>           "sd": "0.56",
#>           "p0": "3",
#>           "p25": "3.94",
#>           "p50": "4.25",
#>           "p75": "4.75",
#>           "p100": "5",
#>           "hist": "▃▂▂▁▇▃▇▂",
#>           "@type": "http://formr.org/codebook/SummaryStatistics"
#>         }
#>       ],
#>       "value": "Number",
#>       "@context": {
#>         "@vocab": "http://pending.schema.org/"
#>       },
#>       "@type": "propertyValue"
#>     },
#>     {
#>       "scale_item_names": ["BFIK_consc_3", "BFIK_consc_4", "BFIK_consc_2R", "BFIK_consc_1"],
#>       "name": "BFIK_consc",
#>       "description": "4 BFIK_consc items averaged with aggregation_function",
#>       "data_summary": [
#>         {
#>           "missing": "0",
#>           "complete": "28",
#>           "n": "28",
#>           "mean": "3.65",
#>           "sd": "0.79",
#>           "p0": "2",
#>           "p25": "3.12",
#>           "p50": "3.75",
#>           "p75": "4.25",
#>           "p100": "5",
#>           "hist": "▂▃▁▃▃▇▂▂",
#>           "@type": "http://formr.org/codebook/SummaryStatistics"
#>         }
#>       ],
#>       "value": "Number",
#>       "@context": {
#>         "@vocab": "http://pending.schema.org/"
#>       },
#>       "@type": "propertyValue"
#>     },
#>     {
#>       "scale_item_names": ["BFIK_extra_2", "BFIK_extra_3R", "BFIK_extra_4", "BFIK_extra_1R"],
#>       "name": "BFIK_extra",
#>       "description": "4 BFIK_extra items averaged with aggregation_function",
#>       "data_summary": [
#>         {
#>           "missing": "0",
#>           "complete": "28",
#>           "n": "28",
#>           "mean": "3.85",
#>           "sd": "1.01",
#>           "p0": "1.5",
#>           "p25": "3.25",
#>           "p50": "4.25",
#>           "p75": "4.56",
#>           "p100": "5",
#>           "hist": "▂▁▂▃▁▃▇▇",
#>           "@type": "http://formr.org/codebook/SummaryStatistics"
#>         }
#>       ],
#>       "value": "Number",
#>       "@context": {
#>         "@vocab": "http://pending.schema.org/"
#>       },
#>       "@type": "propertyValue"
#>     },
#>     {
#>       "scale_item_names": ["BFIK_neuro_2R", "BFIK_neuro_3", "BFIK_neuro_4"],
#>       "name": "BFIK_neuro",
#>       "description": "3 BFIK_neuro items averaged with aggregation_function",
#>       "data_summary": [
#>         {
#>           "missing": "0",
#>           "complete": "28",
#>           "n": "28",
#>           "mean": "2.89",
#>           "sd": "0.93",
#>           "p0": "1.33",
#>           "p25": "2",
#>           "p50": "2.83",
#>           "p75": "3.42",
#>           "p100": "4.33",
#>           "hist": "▅▅▂▅▃▅▁▇",
#>           "@type": "http://formr.org/codebook/SummaryStatistics"
#>         }
#>       ],
#>       "value": "Number",
#>       "@context": {
#>         "@vocab": "http://pending.schema.org/"
#>       },
#>       "@type": "propertyValue"
#>     }
#>   ],
#>   "@context": "http://schema.org/",
#>   "@type": "Dataset"
#> }
#> </script>