Codebook metadata table

will generate a table combining metadata from variable attributes with data summaries generated using skimr::skim_to_wide()

codebook_table(results)

Arguments

results
a data frame, ideally with attributes set on variables

Examples

data("bfi")
codebook_table(bfi)

#>             name
#> 1        session
#> 2        created
#> 3       modified
#> 4          ended
#> 5        expired
#> 6    BFIK_open_2
#> 7  BFIK_agree_4R
#> 8   BFIK_extra_2
#> 9  BFIK_agree_1R
#> 10   BFIK_open_1
#> 11 BFIK_neuro_2R
#> 12  BFIK_consc_3
#> 13  BFIK_consc_4
#> 14 BFIK_consc_2R
#> 15 BFIK_agree_3R
#> 16 BFIK_extra_3R
#> 17  BFIK_neuro_3
#> 18  BFIK_neuro_4
#> 19  BFIK_agree_2
#> 20  BFIK_consc_1
#> 21   BFIK_open_4
#> 22  BFIK_extra_4
#> 23 BFIK_extra_1R
#> 24   BFIK_open_3
#> 25    BFIK_agree
#> 26     BFIK_open
#> 27    BFIK_consc
#> 28    BFIK_extra
#> 29    BFIK_neuro
#>                                                                         label
#> 1                                                                        <NA>
#> 2                                                    user first opened survey
#> 3                                                     user last edited survey
#> 4                                                        user finished survey
#> 5                                                                        <NA>
#> 6                       __Ich bin tiefsinnig, denke gerne über Sachen nach.__
#> 7        __Ich kann mich schroff und abweisend anderen gegenüber verhalten.__
#> 8            __Ich bin begeisterungsfähig und kann andere leicht mitreißen.__
#> 9                                  __Ich neige dazu, andere zu kritisieren.__
#> 10                                       __Ich bin vielseitig interessiert.__
#> 11 __Ich bin entspannt, lasse mich durch Stress nicht aus der Ruhe bringen.__
#> 12                                     __Ich bin tüchtig und arbeite flott.__
#> 13                              __Ich mache Pläne und führe sie auch durch.__
#> 14                                    __Ich bin bequem, neige zur Faulheit.__
#> 15                          __Ich kann mich kalt und distanziert verhalten.__
#> 16                               __Ich bin eher der "stille Typ", wortkarg.__
#> 17                                            __Ich mache mir viele Sorgen.__
#> 18                                  __Ich werde leicht nervös und unsicher.__
#> 19  __Ich schenke anderen leicht Vertrauen, glaube an das Gute im Menschen.__
#> 20                                       __Ich erledige Aufgaben gründlich.__
#> 21                   __Ich schätze künstlerische und ästhetische Eindrücke.__
#> 22                                 __Ich gehe aus mir heraus, bin gesellig.__
#> 23                                __Ich bin eher zurückhaltend, reserviert.__
#> 24             __Ich habe eine aktive Vorstellungskraft, bin phantasievoll.__
#> 25                      4 BFIK_agree items averaged with aggregation_function
#> 26                       4 BFIK_open items averaged with aggregation_function
#> 27                      4 BFIK_consc items averaged with aggregation_function
#> 28                      4 BFIK_extra items averaged with aggregation_function
#> 29                      3 BFIK_neuro items averaged with aggregation_function
#>             type type_options data_type
#> 1           <NA>           NA character
#> 2           <NA>           NA   POSIXct
#> 3           <NA>           NA   POSIXct
#> 4           <NA>           NA   POSIXct
#> 5           <NA>           NA   logical
#> 6  rating_button            5   numeric
#> 7  rating_button            5   numeric
#> 8  rating_button            5   numeric
#> 9  rating_button            5   numeric
#> 10 rating_button            5   numeric
#> 11 rating_button            5   numeric
#> 12 rating_button            5   numeric
#> 13 rating_button            5   numeric
#> 14 rating_button            5   numeric
#> 15 rating_button            5   numeric
#> 16 rating_button            5   numeric
#> 17 rating_button            5   numeric
#> 18 rating_button            5   numeric
#> 19 rating_button            5   numeric
#> 20 rating_button            5   numeric
#> 21 rating_button            5   numeric
#> 22 rating_button            5   numeric
#> 23 rating_button            5   numeric
#> 24 rating_button            5   numeric
#> 25          <NA>           NA   numeric
#> 26          <NA>           NA   numeric
#> 27          <NA>           NA   numeric
#> 28          <NA>           NA   numeric
#> 29          <NA>           NA   numeric
#>                                                                                                                         value_labels
#> 1                                                                                                                               <NA>
#> 2                                                                                                                               <NA>
#> 3                                                                                                                               <NA>
#> 4                                                                                                                               <NA>
#> 5                                                                                                                               <NA>
#> 6  1. 1: Trifft überhaupt nicht zu,\n2. 2,\n3. 3,\n4. 4,\n5. 5: Trifft voll und ganz zu,\nNA. Item was never rendered for this user.
#> 7  5. 1: Trifft überhaupt nicht zu,\n4. 2,\n3. 3,\n2. 4,\n1. 5: Trifft voll und ganz zu,\nNA. Item was never rendered for this user.
#> 8  1. 1: Trifft überhaupt nicht zu,\n2. 2,\n3. 3,\n4. 4,\n5. 5: Trifft voll und ganz zu,\nNA. Item was never rendered for this user.
#> 9  5. 1: Trifft überhaupt nicht zu,\n4. 2,\n3. 3,\n2. 4,\n1. 5: Trifft voll und ganz zu,\nNA. Item was never rendered for this user.
#> 10 1. 1: Trifft überhaupt nicht zu,\n2. 2,\n3. 3,\n4. 4,\n5. 5: Trifft voll und ganz zu,\nNA. Item was never rendered for this user.
#> 11 5. 1: Trifft überhaupt nicht zu,\n4. 2,\n3. 3,\n2. 4,\n1. 5: Trifft voll und ganz zu,\nNA. Item was never rendered for this user.
#> 12 1. 1: Trifft überhaupt nicht zu,\n2. 2,\n3. 3,\n4. 4,\n5. 5: Trifft voll und ganz zu,\nNA. Item was never rendered for this user.
#> 13 1. 1: Trifft überhaupt nicht zu,\n2. 2,\n3. 3,\n4. 4,\n5. 5: Trifft voll und ganz zu,\nNA. Item was never rendered for this user.
#> 14 5. 1: Trifft überhaupt nicht zu,\n4. 2,\n3. 3,\n2. 4,\n1. 5: Trifft voll und ganz zu,\nNA. Item was never rendered for this user.
#> 15 5. 1: Trifft überhaupt nicht zu,\n4. 2,\n3. 3,\n2. 4,\n1. 5: Trifft voll und ganz zu,\nNA. Item was never rendered for this user.
#> 16 5. 1: Trifft überhaupt nicht zu,\n4. 2,\n3. 3,\n2. 4,\n1. 5: Trifft voll und ganz zu,\nNA. Item was never rendered for this user.
#> 17 1. 1: Trifft überhaupt nicht zu,\n2. 2,\n3. 3,\n4. 4,\n5. 5: Trifft voll und ganz zu,\nNA. Item was never rendered for this user.
#> 18 1. 1: Trifft überhaupt nicht zu,\n2. 2,\n3. 3,\n4. 4,\n5. 5: Trifft voll und ganz zu,\nNA. Item was never rendered for this user.
#> 19 1. 1: Trifft überhaupt nicht zu,\n2. 2,\n3. 3,\n4. 4,\n5. 5: Trifft voll und ganz zu,\nNA. Item was never rendered for this user.
#> 20 1. 1: Trifft überhaupt nicht zu,\n2. 2,\n3. 3,\n4. 4,\n5. 5: Trifft voll und ganz zu,\nNA. Item was never rendered for this user.
#> 21 1. 1: Trifft überhaupt nicht zu,\n2. 2,\n3. 3,\n4. 4,\n5. 5: Trifft voll und ganz zu,\nNA. Item was never rendered for this user.
#> 22 1. 1: Trifft überhaupt nicht zu,\n2. 2,\n3. 3,\n4. 4,\n5. 5: Trifft voll und ganz zu,\nNA. Item was never rendered for this user.
#> 23 5. 1: Trifft überhaupt nicht zu,\n4. 2,\n3. 3,\n2. 4,\n1. 5: Trifft voll und ganz zu,\nNA. Item was never rendered for this user.
#> 24 1. 1: Trifft überhaupt nicht zu,\n2. 2,\n3. 3,\n4. 4,\n5. 5: Trifft voll und ganz zu,\nNA. Item was never rendered for this user.
#> 25                                                                                                                              <NA>
#> 26                                                                                                                              <NA>
#> 27                                                                                                                              <NA>
#> 28                                                                                                                              <NA>
#> 29                                                                                                                              <NA>
#>    optional                                          scale_item_names
#> 1        NA                                                      <NA>
#> 2        NA                                                      <NA>
#> 3        NA                                                      <NA>
#> 4        NA                                                      <NA>
#> 5        NA                                                      <NA>
#> 6         0                                                      <NA>
#> 7         0                                                      <NA>
#> 8         0                                                      <NA>
#> 9         0                                                      <NA>
#> 10        0                                                      <NA>
#> 11        0                                                      <NA>
#> 12        0                                                      <NA>
#> 13        0                                                      <NA>
#> 14        0                                                      <NA>
#> 15        0                                                      <NA>
#> 16        0                                                      <NA>
#> 17        0                                                      <NA>
#> 18        0                                                      <NA>
#> 19        0                                                      <NA>
#> 20        0                                                      <NA>
#> 21        0                                                      <NA>
#> 22        0                                                      <NA>
#> 23        0                                                      <NA>
#> 24        0                                                      <NA>
#> 25       NA BFIK_agree_4R, BFIK_agree_1R, BFIK_agree_3R, BFIK_agree_2
#> 26       NA        BFIK_open_2, BFIK_open_1, BFIK_open_4, BFIK_open_3
#> 27       NA   BFIK_consc_3, BFIK_consc_4, BFIK_consc_2R, BFIK_consc_1
#> 28       NA  BFIK_extra_2, BFIK_extra_3R, BFIK_extra_4, BFIK_extra_1R
#> 29       NA                 BFIK_neuro_2R, BFIK_neuro_3, BFIK_neuro_4
#>    item_order missing complete  n empty n_unique count     median        min
#> 1          NA       0       28 28     0       28  <NA>       <NA>         64
#> 2          NA       0       28 28  <NA>       28  <NA> 2016-07-08 2016-07-08
#> 3          NA       0       28 28  <NA>       28  <NA> 2016-07-08 2016-07-08
#> 4          NA       0       28 28  <NA>       28  <NA> 2016-07-08 2016-07-08
#> 5          NA      28        0 28  <NA>     <NA>    28       <NA>       <NA>
#> 6           4       0       28 28  <NA>     <NA>  <NA>       <NA>       <NA>
#> 7           5       0       28 28  <NA>     <NA>  <NA>       <NA>       <NA>
#> 8           6       0       28 28  <NA>     <NA>  <NA>       <NA>       <NA>
#> 9           7       0       28 28  <NA>     <NA>  <NA>       <NA>       <NA>
#> 10          8       0       28 28  <NA>     <NA>  <NA>       <NA>       <NA>
#> 11          9       0       28 28  <NA>     <NA>  <NA>       <NA>       <NA>
#> 12         10       0       28 28  <NA>     <NA>  <NA>       <NA>       <NA>
#> 13         11       0       28 28  <NA>     <NA>  <NA>       <NA>       <NA>
#> 14         12       0       28 28  <NA>     <NA>  <NA>       <NA>       <NA>
#> 15         13       0       28 28  <NA>     <NA>  <NA>       <NA>       <NA>
#> 16         14       0       28 28  <NA>     <NA>  <NA>       <NA>       <NA>
#> 17         15       0       28 28  <NA>     <NA>  <NA>       <NA>       <NA>
#> 18         16       0       28 28  <NA>     <NA>  <NA>       <NA>       <NA>
#> 19         17       0       28 28  <NA>     <NA>  <NA>       <NA>       <NA>
#> 20         18       0       28 28  <NA>     <NA>  <NA>       <NA>       <NA>
#> 21         19       0       28 28  <NA>     <NA>  <NA>       <NA>       <NA>
#> 22         20       0       28 28  <NA>     <NA>  <NA>       <NA>       <NA>
#> 23         21       0       28 28  <NA>     <NA>  <NA>       <NA>       <NA>
#> 24         22       0       28 28  <NA>     <NA>  <NA>       <NA>       <NA>
#> 25         NA       0       28 28  <NA>     <NA>  <NA>       <NA>       <NA>
#> 26         NA       0       28 28  <NA>     <NA>  <NA>       <NA>       <NA>
#> 27         NA       0       28 28  <NA>     <NA>  <NA>       <NA>       <NA>
#> 28         NA       0       28 28  <NA>     <NA>  <NA>       <NA>       <NA>
#> 29         NA       0       28 28  <NA>     <NA>  <NA>       <NA>       <NA>
#>           max mean   sd   p0  p25  p50  p75 p100     hist
#> 1          64 <NA> <NA> <NA> <NA> <NA> <NA> <NA>     <NA>
#> 2  2016-11-02 <NA> <NA> <NA> <NA> <NA> <NA> <NA>     <NA>
#> 3  2016-11-02 <NA> <NA> <NA> <NA> <NA> <NA> <NA>     <NA>
#> 4  2016-11-02 <NA> <NA> <NA> <NA> <NA> <NA> <NA>     <NA>
#> 5        <NA>  NaN <NA> <NA> <NA> <NA> <NA> <NA>     <NA>
#> 6        <NA> 4.21 0.74 2    4    4    5    5    ▁▁▁▁▁▇▁▅
#> 7        <NA> 2.93 1.18 1    2    3    4    5    ▂▇▁▃▁▅▁▂
#> 8        <NA> 4.18 1.09 1    4    4    5    5    ▁▁▁▁▁▇▁▇
#> 9        <NA> 3    0.94 2    2    3    4    5    ▇▁▅▁▁▆▁▁
#> 10       <NA> 4.39 0.83 2    4    5    5    5    ▁▁▂▁▁▃▁▇
#> 11       <NA> 3.11 0.88 2    2    3    4    5    ▆▁▇▁▁▇▁▁
#> 12       <NA> 3.5  1.04 1    3    4    4    5    ▁▂▁▅▁▇▁▂
#> 13       <NA> 3.86 0.76 2    3    4    4    5    ▁▁▃▁▁▇▁▂
#> 14       <NA> 3.18 1.31 1    2    4    4    5    ▃▂▁▃▁▇▁▂
#> 15       <NA> 3.04 1.29 1    2    3    4    5    ▂▇▁▃▁▇▁▃
#> 16       <NA> 3.75 1.21 1    3    4    5    5    ▂▂▁▅▁▇▁▇
#> 17       <NA> 3.07 1.27 1    2    3    4    5    ▃▇▁▇▁▅▁▅
#> 18       <NA> 2.5  1.2  1    1.75 2    4    4    ▆▁▇▁▁▂▁▇
#> 19       <NA> 3.5  1.26 1    2.75 4    4.25 5    ▂▅▁▅▁▇▁▆
#> 20       <NA> 4.07 0.9  2    4    4    5    5    ▁▁▂▁▁▇▁▇
#> 21       <NA> 4.21 0.96 1    4    4    5    5    ▁▁▁▂▁▆▁▇
#> 22       <NA> 3.86 1.11 1    3    4    5    5    ▁▂▁▃▁▇▁▆
#> 23       <NA> 3.61 1.2  1    3    4    5    5    ▁▅▁▆▁▇▁▇
#> 24       <NA> 4.21 0.96 2    4    4.5  5    5    ▁▁▂▁▁▅▁▇
#> 25       <NA> 3.12 0.93 1.5  2.25 3    4    4.75 ▂▇▂▇▃▁▅▅
#> 26       <NA> 4.26 0.56 3    3.94 4.25 4.75 5    ▃▂▂▁▇▃▇▂
#> 27       <NA> 3.65 0.79 2    3.12 3.75 4.25 5    ▂▃▁▃▃▇▂▂
#> 28       <NA> 3.85 1.01 1.5  3.25 4.25 4.56 5    ▂▁▂▃▁▃▇▇
#> 29       <NA> 2.89 0.93 1.33 2    2.83 3.42 4.33 ▅▅▂▅▃▅▁▇