Detect item scales
Source: R/correct_attributes.R

Did you create aggregates of items like this scale <- scale_1 + scale_2R + scale_3R? If you run this function on a dataset, it will detect these relationships and set the appropriate attributes. Once they are set, the codebook package can perform reliability computations for you.

detect_scales(data, quiet = FALSE)

Arguments

data
the data frame

quiet
defaults to false. Suppresses messages about found items.

Examples

bfi <- data.frame(matrix(data = rnorm(500), ncol = 5))
names(bfi) <- c("bfi_e1", "bfi_e2R", "bfi_e3", "bfi_n1", "bfi_n2")
bfi$bfi_e <- rowMeans(bfi[, c("bfi_e1", "bfi_e2R", "bfi_e3")])
bfi <- detect_scales(bfi)

#> 3 bfi_e items connected to scale

#> Warning: bfi_n items found, but no aggregate

bfi$bfi_e

#>   [1]  0.80155  0.12420  1.41116 -1.23598  0.87552  0.97043  0.12567  0.45798
#>   [9]  0.14812 -0.26136 -0.04542  1.28478  0.09695 -0.46463  0.98728  1.06074
#>  [17]  0.80924 -0.39091 -0.16315 -0.01683  0.03796 -1.33485 -0.34182 -0.34697
#>  [25]  0.26811  0.01605  0.67959 -0.63407 -0.96984  0.53467 -1.17722  0.50136
#>  [33]  0.74803  0.72101 -0.50860 -0.75308  0.73761  0.78366  0.23926  0.96985
#>  [41]  0.20763  0.41187 -0.07350  0.45064  0.21954  0.18026 -0.22742  0.04779
#>  [49] -0.81539  0.67904 -1.01231 -0.19294  0.36038 -0.94525 -0.24757  0.71944
#>  [57] -0.25543 -0.43328  0.85900  0.95526  0.19245 -0.53120  1.46977 -0.39889
#>  [65] -0.00019 -0.39563  0.81604 -0.32667 -0.17564 -0.60173  0.14128  1.31901
#>  [73]  0.45576 -0.01858  0.33023 -0.12449  0.46626 -0.69807  0.18845  0.16457
#>  [81]  0.61810 -0.93400 -0.63599 -0.31765 -0.47883  0.23733  0.19460  0.04560
#>  [89] -1.28336  0.06412  0.03258  0.14601  1.35534  0.69666 -0.34551  0.10937
#>  [97]  1.04341  0.10109 -0.80422 -0.40350
#> attr(,"scale_item_names")
#> [1] "bfi_e1"  "bfi_e2R" "bfi_e3" 
#> attr(,"label")
#> [1] "aggregate of 3 bfi_e items"