Detect item scales

Did you create aggregates of items like this scale <- scale_1 + scale_2R + scale_3R? If you run this function on a dataset, it will detect these relationships and set the appropriate attributes. Once they are set, the codebook package can perform reliability computations for you.

Did you create aggregates of items like this scale <- scale_1 + scale_2R + scale_3R? If you run this function on a dataset, it will detect these relationships and set the appropriate attributes. Once they are set, the codebook package can perform reliability computations for you.

Usage

detect_scales(data, quiet = FALSE)

detect_scales(data, quiet = FALSE)

Arguments

data

the data frame

quiet

defaults to false. Suppresses messages about found items.

Examples

bfi <- data.frame(matrix(data = rnorm(300), ncol = 3))
names(bfi) <- c("bfi_e1", "bfi_e2R", "bfi_e3")
bfi$bfi_e <- rowMeans(bfi[, c("bfi_e1", "bfi_e2R", "bfi_e3")])
bfi <- detect_scales(bfi)
#> 3 bfi_e items connected to scale
bfi$bfi_e
#>   [1]  0.236808444 -0.476231621  0.457406176  0.463877128  0.894566061
#>   [6] -0.735737980  0.584123087 -0.765629486  0.839102801  0.832257782
#>  [11]  0.628080093  0.305613576 -0.091518479  0.740314230 -0.669431227
#>  [16]  0.219725369  0.221648528  1.352033602  0.341207131 -0.322339264
#>  [21]  0.731580798  0.316880552  0.280185928 -0.276101859 -0.645748984
#>  [26] -0.904615946  0.241353592  0.107527220 -0.827552380  0.259324663
#>  [31] -0.312794821  0.096077484 -1.591635428 -0.128159412  0.620630736
#>  [36]  0.745208529  0.040945937  0.599525481 -0.086424812  0.818545114
#>  [41]  0.776709916 -0.065373621 -0.375101755  0.679923868  0.528719668
#>  [46] -1.035688752 -1.303638109  0.198283203  0.428955035 -0.130622544
#>  [51] -0.320518057  0.432429263 -0.065524908 -0.857287452 -0.590512927
#>  [56]  1.289659805 -0.519386656 -1.326294682  0.346165460 -0.075591161
#>  [61]  0.583690346 -0.919761838 -0.130995173 -0.675317248  0.393645220
#>  [66] -0.044075001 -0.648013090 -0.157806380 -0.245149660 -1.122878115
#>  [71]  0.462284564 -1.167456982 -0.188661742  1.344595739  0.307784430
#>  [76]  0.739960072  0.728898483 -0.435223051 -0.029625667  0.129053126
#>  [81] -0.764220076  0.439199573 -0.238284862 -0.450978705  0.329856393
#>  [86]  1.309939340 -1.310376849  0.590375873  0.245267591 -1.252435901
#>  [91] -0.875119070 -0.002765186  0.099447534 -0.236540099  0.662425478
#>  [96] -1.546541499 -0.494083596  0.523752020 -0.772030576  0.113826350
#> attr(,"scale_item_names")
#> [1] "bfi_e1"  "bfi_e2R" "bfi_e3" 
#> attr(,"label")
#> [1] "aggregate of 3 bfi_e items"
bfi <- data.frame(matrix(data = rnorm(300), ncol = 3))
names(bfi) <- c("bfi_e1", "bfi_e2R", "bfi_e3")
bfi$bfi_e <- rowMeans(bfi[, c("bfi_e1", "bfi_e2R", "bfi_e3")])
bfi <- detect_scales(bfi)
#> 3 bfi_e items connected to scale
bfi$bfi_e
#>   [1]  0.256596110  0.661218543 -0.664322719  0.434891800 -0.031779575
#>   [6]  0.365696596 -1.306240212 -0.623121950 -0.612711377 -1.106205100
#>  [11]  0.614808891 -0.075827088  0.182669487 -0.451533900 -0.454538313
#>  [16]  0.910818570 -0.405420594  0.233886875 -0.937625452 -0.193737168
#>  [21] -0.228149409  0.687654432  0.395421746  0.599151407  1.174867760
#>  [26] -0.155523117  1.098220036 -0.456091906 -0.336469935  1.512587955
#>  [31]  0.399805250 -0.396619665 -0.269658597 -0.214700479 -0.378173689
#>  [36]  0.443365702  0.908733475  0.052400284  0.242154948  0.149008884
#>  [41]  0.209647490  0.356641492  0.614080058 -0.274160930 -0.084715522
#>  [46] -0.531040879 -0.422708294 -0.470475155 -0.814858208 -0.588315100
#>  [51]  1.091467536  0.320599935  0.996773978 -0.200175736 -0.446062843
#>  [56]  0.543623739 -0.167868130 -0.007257087  0.656378028  1.005426062
#>  [61] -0.560318383  0.133006507 -0.188995550 -1.334626037  1.470283416
#>  [66] -0.527778950  0.287272734 -0.306084078 -0.312422337 -0.129541534
#>  [71] -0.525578583 -0.759719946 -1.071909603  0.601158262 -0.297178907
#>  [76] -0.165597041  0.777162680  1.383476210 -0.256776664  0.041581469
#>  [81] -0.822791525  0.586172987 -0.522463683  0.723380995 -0.679383729
#>  [86]  0.379866122 -0.806330625  0.018557656 -0.199364879  0.515841394
#>  [91] -0.579345474  0.769985253 -0.514024391  0.200019489  0.183214276
#>  [96] -0.908284937  0.360489375  0.362312243 -0.261189347  0.753366895
#> attr(,"scale_item_names")
#> [1] "bfi_e1"  "bfi_e2R" "bfi_e3" 
#> attr(,"label")
#> [1] "aggregate of 3 bfi_e items"