detect_scales

Did you create aggregates of items like this scale = scale_1 + scale_2R + scale_3R ? If you run this function on a dataset, it will detect these relationships and set the appropriate attributes. Once they are set, the codebook package can perform reliability computations for you.

detect_scales(data)

Arguments

data
the data frame

Examples

bfi = data.frame(matrix(data = rnorm(500), ncol = 5))
names(bfi) = c("bfi_e1", "bfi_e2R", "bfi_e3", "bfi_n1", "bfi_n2")
bfi$bfi_e = rowMeans(bfi[, c("bfi_e1", "bfi_e2R", "bfi_e3")])
bfi = detect_scales(bfi)

#> 3 bfi_e items found and connected to scale

#> Warning: bfi_n items found, but no aggregate

bfi$bfi_e

#>   [1] -0.738879060  0.276781909 -1.188285970 -0.357755467 -1.193800912
#>   [6]  0.174453450 -1.159622852 -0.472388935 -0.693575642  1.114893833
#>  [11] -0.638660140 -0.370355431  1.036641587 -1.114433988  0.541056120
#>  [16]  0.112981622  0.153786783 -0.440492348 -0.219967214  0.366656515
#>  [21]  0.722862604  0.245610262  0.042491631  0.429547978  0.332552124
#>  [26]  0.386079898  0.002729382 -0.250434455 -0.215433325 -0.949736315
#>  [31]  0.008862955 -0.096713657  0.357477894 -0.414695520 -0.253940206
#>  [36] -0.106396730 -1.203495236  0.036370746  0.578297147 -0.183964141
#>  [41] -0.331818812 -0.049599932  0.975576822  0.572908198 -0.002172322
#>  [46]  1.820036476 -0.356684503  0.338140688 -0.316492245 -0.083352038
#>  [51]  0.153673198 -0.208242078 -0.845913592 -0.430186953  0.845980231
#>  [56]  0.132969218 -0.648613616  0.615374171 -0.061068836  0.825959182
#>  [61] -0.225345196 -0.387541688  0.264865866 -0.065021923 -1.291498448
#>  [66] -1.288753980  0.608465372 -0.157600878  0.033379286 -0.313940621
#>  [71] -0.504248694 -0.955831759  0.830044190  0.436618619  0.698300317
#>  [76]  1.145861477  0.763872509 -0.225185342  0.211181473  0.604410260
#>  [81]  0.631558936 -0.255727563 -0.121051986  0.801548578  0.124200871
#>  [86]  1.411160104 -1.235984383  0.875517402  0.970433743  0.125670240
#>  [91]  0.457980015  0.148118662 -0.261359084 -0.045420751  1.284781875
#>  [96]  0.096950076 -0.464628583  0.987277622  1.060740615  0.809242736
#> attr(,"scale_item_names")
#> [1] "bfi_e1"  "bfi_e2R" "bfi_e3" 
#> attr(,"label")
#> [1] "aggregate of 3 bfi_e items"