Detect item scales

Did you create aggregates of items like this scale <- scale_1 + scale_2R + scale_3R? If you run this function on a dataset, it will detect these relationships and set the appropriate attributes. Once they are set, the codebook package can perform reliability computations for you.

detect_scales(data, quiet = FALSE)

Arguments

data
the data frame

quiet
defaults to false. Suppresses messages about found items.

Examples

bfi <- data.frame(matrix(data = rnorm(500), ncol = 5))
names(bfi) <- c("bfi_e1", "bfi_e2R", "bfi_e3", "bfi_n1", "bfi_n2")
bfi$bfi_e <- rowMeans(bfi[, c("bfi_e1", "bfi_e2R", "bfi_e3")])
bfi <- detect_scales(bfi)

#> 3 bfi_e items connected to scale

#> Warning: bfi_n items found, but no aggregate

bfi$bfi_e

#>   [1]  0.1924536606 -0.5312016165  1.4697668632 -0.3988861741 -0.0001894151
#>   [6] -0.3956289014  0.8160424138 -0.3266707117 -0.1756387935 -0.6017318640
#>  [11]  0.1412798858  1.3190123866  0.4557577536 -0.0185777490  0.3302301687
#>  [16] -0.1244923341  0.4662560852 -0.6980709833  0.1884465481  0.1645677940
#>  [21]  0.6181046675 -0.9340002428 -0.6359859871 -0.3176546090 -0.4788309317
#>  [26]  0.2373336437  0.1946028667  0.0456007471 -1.2833583570  0.0641209995
#>  [31]  0.0325788440  0.1460103547  1.3553405723  0.6966568876 -0.3455142832
#>  [36]  0.1093705621  1.0434134739  0.1010924876 -0.8042213045 -0.4034993055
#>  [41]  0.6277291548  0.4356623000  1.0479255488 -0.3626773462  1.2376966704
#>  [46]  0.5975538321  0.3654544478  0.1792049220  0.1051717067  0.0025089560
#>  [51] -0.2676596739  1.1255690061  0.2173858347 -0.7937368848  1.0581547208
#>  [56]  0.5540621767  0.7821132260  0.3168688466 -0.1830835659  0.0148958715
#>  [61]  0.6423043116 -1.2481138922 -0.5065791126  0.0492932507  1.1474103621
#>  [66] -0.2642429787 -0.1914422633 -0.5679208891 -0.7222018261  0.2253117078
#>  [71] -1.1894831281  0.7516472794  0.2512675891  0.7508354601 -0.5309632969
#>  [76] -0.5383461161  0.4543405664  0.2974714244  0.0301725562  0.7529011373
#>  [81] -0.2670387714 -0.6036551925  0.1587362734  0.0167693154 -0.3242835091
#>  [86]  0.3817768686  0.0003000125  0.5811360532 -1.3810287593  0.8405386016
#>  [91] -0.1931955214 -0.1300380884  0.0458491608 -1.0488637773 -0.3118260167
#>  [96]  0.4595724652 -0.8672159132 -0.8482639098  1.3711669087 -0.0838575223
#> attr(,"scale_item_names")
#> [1] "bfi_e1"  "bfi_e2R" "bfi_e3" 
#> attr(,"label")
#> [1] "aggregate of 3 bfi_e items"