Metadata as JSON-LD

Echo a list of a metadata, generated using metadata_list() as JSON-LD in a script tag.

metadata_jsonld(results)

Arguments

results
a data frame, ideally with attributes set on variables

Examples

data("bfi")
metadata_jsonld(bfi)

#> 
#> 
#> 
#> <script type="application/ld+json">
#> {
#>   "@context": "http://schema.org/",
#>   "@type": "Dataset",
#>   "variableMeasured": [
#>     {
#>       "name": "A1",
#>       "@type": "propertyValue"
#>     },
#>     {
#>       "name": "A2",
#>       "@type": "propertyValue"
#>     },
#>     {
#>       "name": "A3",
#>       "@type": "propertyValue"
#>     },
#>     {
#>       "name": "A4",
#>       "@type": "propertyValue"
#>     },
#>     {
#>       "name": "A5",
#>       "@type": "propertyValue"
#>     },
#>     {
#>       "name": "C1",
#>       "@type": "propertyValue"
#>     },
#>     {
#>       "name": "C2",
#>       "@type": "propertyValue"
#>     },
#>     {
#>       "name": "C3",
#>       "@type": "propertyValue"
#>     },
#>     {
#>       "name": "C4",
#>       "@type": "propertyValue"
#>     },
#>     {
#>       "name": "C5",
#>       "@type": "propertyValue"
#>     },
#>     {
#>       "name": "E1",
#>       "@type": "propertyValue"
#>     },
#>     {
#>       "name": "E2",
#>       "@type": "propertyValue"
#>     },
#>     {
#>       "name": "E3",
#>       "@type": "propertyValue"
#>     },
#>     {
#>       "name": "E4",
#>       "@type": "propertyValue"
#>     },
#>     {
#>       "name": "E5",
#>       "@type": "propertyValue"
#>     },
#>     {
#>       "name": "N1",
#>       "@type": "propertyValue"
#>     },
#>     {
#>       "name": "N2",
#>       "@type": "propertyValue"
#>     },
#>     {
#>       "name": "N3",
#>       "@type": "propertyValue"
#>     },
#>     {
#>       "name": "N4",
#>       "@type": "propertyValue"
#>     },
#>     {
#>       "name": "N5",
#>       "@type": "propertyValue"
#>     },
#>     {
#>       "name": "O1",
#>       "@type": "propertyValue"
#>     },
#>     {
#>       "name": "O2",
#>       "@type": "propertyValue"
#>     },
#>     {
#>       "name": "O3",
#>       "@type": "propertyValue"
#>     },
#>     {
#>       "name": "O4",
#>       "@type": "propertyValue"
#>     },
#>     {
#>       "name": "O5",
#>       "@type": "propertyValue"
#>     },
#>     {
#>       "name": "gender",
#>       "@type": "propertyValue"
#>     },
#>     {
#>       "name": "education",
#>       "@type": "propertyValue"
#>     },
#>     {
#>       "name": "age",
#>       "@type": "propertyValue"
#>     }
#>   ],
#>   "description": "\n\n\n## Table of variables\nThis table contains variable names, labels, their central tendencies and other attributes.\n\n|name      |data_type | n_missing| complete_rate|min |median |max |      mean|         sd|hist  |\n|:---------|:---------|---------:|-------------:|:---|:------|:---|---------:|----------:|:-----|\n|A1        |numeric   |        16|     0.9942857|1   |2      |6   |  2.413434|  1.4077372|▇▂▂▁▁ |\n|A2        |numeric   |        27|     0.9903571|1   |5      |6   |  4.802380|  1.1720199|▂▁▅▇▇ |\n|A3        |numeric   |        26|     0.9907143|1   |5      |6   |  4.603821|  1.3018336|▂▂▅▇▆ |\n|A4        |numeric   |        19|     0.9932143|1   |5      |6   |  4.699748|  1.4796327|▂▁▃▅▇ |\n|A5        |numeric   |        16|     0.9942857|1   |5      |6   |  4.560345|  1.2585121|▂▂▅▇▆ |\n|C1        |numeric   |        21|     0.9925000|1   |5      |6   |  4.502339|  1.2413465|▂▂▅▇▅ |\n|C2        |numeric   |        24|     0.9914286|1   |5      |6   |  4.369957|  1.3183465|▃▂▆▇▅ |\n|C3        |numeric   |        20|     0.9928571|1   |5      |6   |  4.303957|  1.2885518|▃▂▆▇▅ |\n|C4        |numeric   |        26|     0.9907143|1   |2      |6   |  2.553353|  1.3751181|▇▂▂▁▁ |\n|C5        |numeric   |        16|     0.9942857|1   |3      |6   |  3.296695|  1.6285419|▇▂▅▃▂ |\n|E1        |numeric   |        23|     0.9917857|1   |3      |6   |  2.974433|  1.6315055|▇▂▃▂▂ |\n|E2        |numeric   |        16|     0.9942857|1   |3      |6   |  3.141882|  1.6052103|▇▂▃▂▂ |\n|E3        |numeric   |        25|     0.9910714|1   |4      |6   |  4.000721|  1.3527188|▅▃▇▇▃ |\n|E4        |numeric   |         9|     0.9967857|1   |5      |6   |  4.422429|  1.4575174|▃▂▃▇▆ |\n|E5        |numeric   |        21|     0.9925000|1   |5      |6   |  4.416337|  1.3347677|▃▂▅▇▅ |\n|N1        |numeric   |        22|     0.9921429|1   |3      |6   |  2.929086|  1.5709175|▇▂▃▂▁ |\n|N2        |numeric   |        21|     0.9925000|1   |4      |6   |  3.507737|  1.5259436|▇▃▆▅▃ |\n|N3        |numeric   |        11|     0.9960714|1   |3      |6   |  3.216565|  1.6029021|▇▂▅▃▂ |\n|N4        |numeric   |        36|     0.9871429|1   |3      |6   |  3.185601|  1.5696851|▇▃▅▃▂ |\n|N5        |numeric   |        29|     0.9896429|1   |3      |6   |  2.969686|  1.6186474|▇▂▃▂▂ |\n|O1        |numeric   |        22|     0.9921429|1   |5      |6   |  4.816055|  1.1295303|▁▂▅▇▇ |\n|O2        |numeric   |         0|     1.0000000|1   |2      |6   |  2.713214|  1.5651518|▇▂▂▂▁ |\n|O3        |numeric   |        28|     0.9900000|1   |5      |6   |  4.438312|  1.2209011|▂▂▆▇▅ |\n|O4        |numeric   |        14|     0.9950000|1   |5      |6   |  4.892319|  1.2212504|▁▁▃▆▇ |\n|O5        |numeric   |        20|     0.9928571|1   |2      |6   |  2.489568|  1.3279590|▇▂▂▁▁ |\n|gender    |numeric   |         0|     1.0000000|1   |2      |2   |  1.671786|  0.4696471|▃▁▁▁▇ |\n|education |numeric   |       223|     0.9203571|1   |3      |5   |  3.190144|  1.1077139|▂▂▇▂▃ |\n|age       |numeric   |         0|     1.0000000|3   |26     |86  | 28.782143| 11.1275548|▃▇▂▁▁ |\n\n### Note\nThis dataset was automatically described using the [codebook R package](https://rubenarslan.github.io/codebook/) (version 0.8.2)."
#> }
#> </script>
#> 
#> <details><summary>JSON-LD metadata</summary>
#> The following JSON-LD can be found by search engines, if you share this codebook
#> publicly on the web.
#> 
#> ```json
#> {
#>   "@context": "http://schema.org/",
#>   "@type": "Dataset",
#>   "variableMeasured": [
#>     {
#>       "name": "A1",
#>       "@type": "propertyValue"
#>     },
#>     {
#>       "name": "A2",
#>       "@type": "propertyValue"
#>     },
#>     {
#>       "name": "A3",
#>       "@type": "propertyValue"
#>     },
#>     {
#>       "name": "A4",
#>       "@type": "propertyValue"
#>     },
#>     {
#>       "name": "A5",
#>       "@type": "propertyValue"
#>     },
#>     {
#>       "name": "C1",
#>       "@type": "propertyValue"
#>     },
#>     {
#>       "name": "C2",
#>       "@type": "propertyValue"
#>     },
#>     {
#>       "name": "C3",
#>       "@type": "propertyValue"
#>     },
#>     {
#>       "name": "C4",
#>       "@type": "propertyValue"
#>     },
#>     {
#>       "name": "C5",
#>       "@type": "propertyValue"
#>     },
#>     {
#>       "name": "E1",
#>       "@type": "propertyValue"
#>     },
#>     {
#>       "name": "E2",
#>       "@type": "propertyValue"
#>     },
#>     {
#>       "name": "E3",
#>       "@type": "propertyValue"
#>     },
#>     {
#>       "name": "E4",
#>       "@type": "propertyValue"
#>     },
#>     {
#>       "name": "E5",
#>       "@type": "propertyValue"
#>     },
#>     {
#>       "name": "N1",
#>       "@type": "propertyValue"
#>     },
#>     {
#>       "name": "N2",
#>       "@type": "propertyValue"
#>     },
#>     {
#>       "name": "N3",
#>       "@type": "propertyValue"
#>     },
#>     {
#>       "name": "N4",
#>       "@type": "propertyValue"
#>     },
#>     {
#>       "name": "N5",
#>       "@type": "propertyValue"
#>     },
#>     {
#>       "name": "O1",
#>       "@type": "propertyValue"
#>     },
#>     {
#>       "name": "O2",
#>       "@type": "propertyValue"
#>     },
#>     {
#>       "name": "O3",
#>       "@type": "propertyValue"
#>     },
#>     {
#>       "name": "O4",
#>       "@type": "propertyValue"
#>     },
#>     {
#>       "name": "O5",
#>       "@type": "propertyValue"
#>     },
#>     {
#>       "name": "gender",
#>       "@type": "propertyValue"
#>     },
#>     {
#>       "name": "education",
#>       "@type": "propertyValue"
#>     },
#>     {
#>       "name": "age",
#>       "@type": "propertyValue"
#>     }
#>   ],
#>   "description": "\n\n\n## Table of variables\nThis table contains variable names, labels, their central tendencies and other attributes.\n\n|name      |data_type | n_missing| complete_rate|min |median |max |      mean|         sd|hist  |\n|:---------|:---------|---------:|-------------:|:---|:------|:---|---------:|----------:|:-----|\n|A1        |numeric   |        16|     0.9942857|1   |2      |6   |  2.413434|  1.4077372|▇▂▂▁▁ |\n|A2        |numeric   |        27|     0.9903571|1   |5      |6   |  4.802380|  1.1720199|▂▁▅▇▇ |\n|A3        |numeric   |        26|     0.9907143|1   |5      |6   |  4.603821|  1.3018336|▂▂▅▇▆ |\n|A4        |numeric   |        19|     0.9932143|1   |5      |6   |  4.699748|  1.4796327|▂▁▃▅▇ |\n|A5        |numeric   |        16|     0.9942857|1   |5      |6   |  4.560345|  1.2585121|▂▂▅▇▆ |\n|C1        |numeric   |        21|     0.9925000|1   |5      |6   |  4.502339|  1.2413465|▂▂▅▇▅ |\n|C2        |numeric   |        24|     0.9914286|1   |5      |6   |  4.369957|  1.3183465|▃▂▆▇▅ |\n|C3        |numeric   |        20|     0.9928571|1   |5      |6   |  4.303957|  1.2885518|▃▂▆▇▅ |\n|C4        |numeric   |        26|     0.9907143|1   |2      |6   |  2.553353|  1.3751181|▇▂▂▁▁ |\n|C5        |numeric   |        16|     0.9942857|1   |3      |6   |  3.296695|  1.6285419|▇▂▅▃▂ |\n|E1        |numeric   |        23|     0.9917857|1   |3      |6   |  2.974433|  1.6315055|▇▂▃▂▂ |\n|E2        |numeric   |        16|     0.9942857|1   |3      |6   |  3.141882|  1.6052103|▇▂▃▂▂ |\n|E3        |numeric   |        25|     0.9910714|1   |4      |6   |  4.000721|  1.3527188|▅▃▇▇▃ |\n|E4        |numeric   |         9|     0.9967857|1   |5      |6   |  4.422429|  1.4575174|▃▂▃▇▆ |\n|E5        |numeric   |        21|     0.9925000|1   |5      |6   |  4.416337|  1.3347677|▃▂▅▇▅ |\n|N1        |numeric   |        22|     0.9921429|1   |3      |6   |  2.929086|  1.5709175|▇▂▃▂▁ |\n|N2        |numeric   |        21|     0.9925000|1   |4      |6   |  3.507737|  1.5259436|▇▃▆▅▃ |\n|N3        |numeric   |        11|     0.9960714|1   |3      |6   |  3.216565|  1.6029021|▇▂▅▃▂ |\n|N4        |numeric   |        36|     0.9871429|1   |3      |6   |  3.185601|  1.5696851|▇▃▅▃▂ |\n|N5        |numeric   |        29|     0.9896429|1   |3      |6   |  2.969686|  1.6186474|▇▂▃▂▂ |\n|O1        |numeric   |        22|     0.9921429|1   |5      |6   |  4.816055|  1.1295303|▁▂▅▇▇ |\n|O2        |numeric   |         0|     1.0000000|1   |2      |6   |  2.713214|  1.5651518|▇▂▂▂▁ |\n|O3        |numeric   |        28|     0.9900000|1   |5      |6   |  4.438312|  1.2209011|▂▂▆▇▅ |\n|O4        |numeric   |        14|     0.9950000|1   |5      |6   |  4.892319|  1.2212504|▁▁▃▆▇ |\n|O5        |numeric   |        20|     0.9928571|1   |2      |6   |  2.489568|  1.3279590|▇▂▂▁▁ |\n|gender    |numeric   |         0|     1.0000000|1   |2      |2   |  1.671786|  0.4696471|▃▁▁▁▇ |\n|education |numeric   |       223|     0.9203571|1   |3      |5   |  3.190144|  1.1077139|▂▂▇▂▃ |\n|age       |numeric   |         0|     1.0000000|3   |26     |86  | 28.782143| 11.1275548|▃▇▂▁▁ |\n\n### Note\nThis dataset was automatically described using the [codebook R package](https://rubenarslan.github.io/codebook/) (version 0.8.2)."
#> }`
#> ```
#> </details>