Metadata as JSON-LD

Echo a list of a metadata, generated using metadata_list() as JSON-LD in a script tag.

metadata_jsonld(results)

Arguments

results
a data frame, ideally with attributes set on variables

Examples

data("bfi")
metadata_jsonld(bfi)

#> 
#> 
#> 
#> <script type="application/ld+json">
#> {
#>   "@context": "http://schema.org/",
#>   "@type": "Dataset",
#>   "variableMeasured": [
#>     {
#>       "name": "A1",
#>       "@type": "propertyValue"
#>     },
#>     {
#>       "name": "A2",
#>       "@type": "propertyValue"
#>     },
#>     {
#>       "name": "A3",
#>       "@type": "propertyValue"
#>     },
#>     {
#>       "name": "A4",
#>       "@type": "propertyValue"
#>     },
#>     {
#>       "name": "A5",
#>       "@type": "propertyValue"
#>     },
#>     {
#>       "name": "C1",
#>       "@type": "propertyValue"
#>     },
#>     {
#>       "name": "C2",
#>       "@type": "propertyValue"
#>     },
#>     {
#>       "name": "C3",
#>       "@type": "propertyValue"
#>     },
#>     {
#>       "name": "C4",
#>       "@type": "propertyValue"
#>     },
#>     {
#>       "name": "C5",
#>       "@type": "propertyValue"
#>     },
#>     {
#>       "name": "E1",
#>       "@type": "propertyValue"
#>     },
#>     {
#>       "name": "E2",
#>       "@type": "propertyValue"
#>     },
#>     {
#>       "name": "E3",
#>       "@type": "propertyValue"
#>     },
#>     {
#>       "name": "E4",
#>       "@type": "propertyValue"
#>     },
#>     {
#>       "name": "E5",
#>       "@type": "propertyValue"
#>     },
#>     {
#>       "name": "N1",
#>       "@type": "propertyValue"
#>     },
#>     {
#>       "name": "N2",
#>       "@type": "propertyValue"
#>     },
#>     {
#>       "name": "N3",
#>       "@type": "propertyValue"
#>     },
#>     {
#>       "name": "N4",
#>       "@type": "propertyValue"
#>     },
#>     {
#>       "name": "N5",
#>       "@type": "propertyValue"
#>     },
#>     {
#>       "name": "O1",
#>       "@type": "propertyValue"
#>     },
#>     {
#>       "name": "O2",
#>       "@type": "propertyValue"
#>     },
#>     {
#>       "name": "O3",
#>       "@type": "propertyValue"
#>     },
#>     {
#>       "name": "O4",
#>       "@type": "propertyValue"
#>     },
#>     {
#>       "name": "O5",
#>       "@type": "propertyValue"
#>     },
#>     {
#>       "name": "gender",
#>       "@type": "propertyValue"
#>     },
#>     {
#>       "name": "education",
#>       "@type": "propertyValue"
#>     },
#>     {
#>       "name": "age",
#>       "@type": "propertyValue"
#>     }
#>   ],
#>   "description": "\n\n\n## Table of variables\nThis table contains variable names, labels, their central tendencies and other attributes.\n\n|name      |data_type |missing |complete |n    |mean  |sd    |p0 |p25 |p50 |p75 |p100 |hist     |\n|:---------|:---------|:-------|:--------|:----|:-----|:-----|:--|:---|:---|:---|:----|:--------|\n|A1        |integer   |16      |2784     |2800 |2.41  |1.41  |1  |1   |2   |3   |6    |▇▇▁▃▃▁▂▁ |\n|A2        |integer   |27      |2773     |2800 |4.8   |1.17  |1  |4   |5   |6   |6    |▁▁▁▁▅▁▇▇ |\n|A3        |integer   |26      |2774     |2800 |4.6   |1.3   |1  |4   |5   |6   |6    |▁▂▁▂▅▁▇▆ |\n|A4        |integer   |19      |2781     |2800 |4.7   |1.48  |1  |4   |5   |6   |6    |▁▂▁▁▃▁▅▇ |\n|A5        |integer   |16      |2784     |2800 |4.56  |1.26  |1  |4   |5   |5   |6    |▁▂▁▂▅▁▇▆ |\n|C1        |integer   |21      |2779     |2800 |4.5   |1.24  |1  |4   |5   |5   |6    |▁▁▁▂▅▁▇▅ |\n|C2        |integer   |24      |2776     |2800 |4.37  |1.32  |1  |4   |5   |5   |6    |▁▂▁▂▆▁▇▅ |\n|C3        |integer   |20      |2780     |2800 |4.3   |1.29  |1  |4   |5   |5   |6    |▁▂▁▂▆▁▇▅ |\n|C4        |integer   |26      |2774     |2800 |2.55  |1.38  |1  |1   |2   |4   |6    |▇▇▁▅▅▁▂▁ |\n|C5        |integer   |16      |2784     |2800 |3.3   |1.63  |1  |2   |3   |5   |6    |▆▇▁▅▇▁▆▃ |\n|E1        |integer   |23      |2777     |2800 |2.97  |1.63  |1  |2   |3   |4   |6    |▇▇▁▅▆▁▅▃ |\n|E2        |integer   |16      |2784     |2800 |3.14  |1.61  |1  |2   |3   |4   |6    |▆▇▁▅▇▁▅▃ |\n|E3        |integer   |25      |2775     |2800 |4     |1.35  |1  |3   |4   |5   |6    |▂▃▁▃▇▁▇▃ |\n|E4        |integer   |9       |2791     |2800 |4.42  |1.46  |1  |4   |5   |6   |6    |▁▂▁▂▃▁▇▆ |\n|E5        |integer   |21      |2779     |2800 |4.42  |1.33  |1  |4   |5   |5   |6    |▁▂▁▂▅▁▇▅ |\n|N1        |integer   |22      |2778     |2800 |2.93  |1.57  |1  |2   |3   |4   |6    |▇▇▁▅▆▁▅▂ |\n|N2        |integer   |21      |2779     |2800 |3.51  |1.53  |1  |2   |4   |5   |6    |▃▆▁▅▇▁▆▃ |\n|N3        |integer   |11      |2789     |2800 |3.22  |1.6   |1  |2   |3   |4   |6    |▆▇▁▅▇▁▆▃ |\n|N4        |integer   |36      |2764     |2800 |3.19  |1.57  |1  |2   |3   |4   |6    |▆▇▁▅▇▁▅▃ |\n|N5        |integer   |29      |2771     |2800 |2.97  |1.62  |1  |2   |3   |4   |6    |▇▇▁▅▆▁▃▃ |\n|O1        |integer   |22      |2778     |2800 |4.82  |1.13  |1  |4   |5   |6   |6    |▁▁▁▂▅▁▇▇ |\n|O2        |integer   |0       |2800     |2800 |2.71  |1.57  |1  |1   |2   |4   |6    |▇▇▁▃▅▁▃▂ |\n|O3        |integer   |28      |2772     |2800 |4.44  |1.22  |1  |4   |5   |5   |6    |▁▁▁▂▆▁▇▅ |\n|O4        |integer   |14      |2786     |2800 |4.89  |1.22  |1  |4   |5   |6   |6    |▁▁▁▁▃▁▆▇ |\n|O5        |integer   |20      |2780     |2800 |2.49  |1.33  |1  |1   |2   |3   |6    |▇▇▁▅▃▁▂▁ |\n|gender    |integer   |0       |2800     |2800 |1.67  |0.47  |1  |1   |2   |2   |2    |▃▁▁▁▁▁▁▇ |\n|education |integer   |223     |2577     |2800 |3.19  |1.11  |1  |3   |3   |4   |5    |▂▂▁▇▁▂▁▃ |\n|age       |integer   |0       |2800     |2800 |28.78 |11.13 |3  |20  |26  |35  |86   |▁▇▆▃▂▁▁▁ |\n\n### Note\nThis dataset was automatically described using the [codebook R package](https://rubenarslan.github.io/codebook/) (version 0.7.6.9000)."
#> }
#> </script>
#> 
#> <details><summary>JSON-LD metadata</summary>
#> The following JSON-LD can be found by search engines, if you share this codebook
#> publicly on the web.
#> 
#> ```json
#> {
#>   "@context": "http://schema.org/",
#>   "@type": "Dataset",
#>   "variableMeasured": [
#>     {
#>       "name": "A1",
#>       "@type": "propertyValue"
#>     },
#>     {
#>       "name": "A2",
#>       "@type": "propertyValue"
#>     },
#>     {
#>       "name": "A3",
#>       "@type": "propertyValue"
#>     },
#>     {
#>       "name": "A4",
#>       "@type": "propertyValue"
#>     },
#>     {
#>       "name": "A5",
#>       "@type": "propertyValue"
#>     },
#>     {
#>       "name": "C1",
#>       "@type": "propertyValue"
#>     },
#>     {
#>       "name": "C2",
#>       "@type": "propertyValue"
#>     },
#>     {
#>       "name": "C3",
#>       "@type": "propertyValue"
#>     },
#>     {
#>       "name": "C4",
#>       "@type": "propertyValue"
#>     },
#>     {
#>       "name": "C5",
#>       "@type": "propertyValue"
#>     },
#>     {
#>       "name": "E1",
#>       "@type": "propertyValue"
#>     },
#>     {
#>       "name": "E2",
#>       "@type": "propertyValue"
#>     },
#>     {
#>       "name": "E3",
#>       "@type": "propertyValue"
#>     },
#>     {
#>       "name": "E4",
#>       "@type": "propertyValue"
#>     },
#>     {
#>       "name": "E5",
#>       "@type": "propertyValue"
#>     },
#>     {
#>       "name": "N1",
#>       "@type": "propertyValue"
#>     },
#>     {
#>       "name": "N2",
#>       "@type": "propertyValue"
#>     },
#>     {
#>       "name": "N3",
#>       "@type": "propertyValue"
#>     },
#>     {
#>       "name": "N4",
#>       "@type": "propertyValue"
#>     },
#>     {
#>       "name": "N5",
#>       "@type": "propertyValue"
#>     },
#>     {
#>       "name": "O1",
#>       "@type": "propertyValue"
#>     },
#>     {
#>       "name": "O2",
#>       "@type": "propertyValue"
#>     },
#>     {
#>       "name": "O3",
#>       "@type": "propertyValue"
#>     },
#>     {
#>       "name": "O4",
#>       "@type": "propertyValue"
#>     },
#>     {
#>       "name": "O5",
#>       "@type": "propertyValue"
#>     },
#>     {
#>       "name": "gender",
#>       "@type": "propertyValue"
#>     },
#>     {
#>       "name": "education",
#>       "@type": "propertyValue"
#>     },
#>     {
#>       "name": "age",
#>       "@type": "propertyValue"
#>     }
#>   ],
#>   "description": "\n\n\n## Table of variables\nThis table contains variable names, labels, their central tendencies and other attributes.\n\n|name      |data_type |missing |complete |n    |mean  |sd    |p0 |p25 |p50 |p75 |p100 |hist     |\n|:---------|:---------|:-------|:--------|:----|:-----|:-----|:--|:---|:---|:---|:----|:--------|\n|A1        |integer   |16      |2784     |2800 |2.41  |1.41  |1  |1   |2   |3   |6    |▇▇▁▃▃▁▂▁ |\n|A2        |integer   |27      |2773     |2800 |4.8   |1.17  |1  |4   |5   |6   |6    |▁▁▁▁▅▁▇▇ |\n|A3        |integer   |26      |2774     |2800 |4.6   |1.3   |1  |4   |5   |6   |6    |▁▂▁▂▅▁▇▆ |\n|A4        |integer   |19      |2781     |2800 |4.7   |1.48  |1  |4   |5   |6   |6    |▁▂▁▁▃▁▅▇ |\n|A5        |integer   |16      |2784     |2800 |4.56  |1.26  |1  |4   |5   |5   |6    |▁▂▁▂▅▁▇▆ |\n|C1        |integer   |21      |2779     |2800 |4.5   |1.24  |1  |4   |5   |5   |6    |▁▁▁▂▅▁▇▅ |\n|C2        |integer   |24      |2776     |2800 |4.37  |1.32  |1  |4   |5   |5   |6    |▁▂▁▂▆▁▇▅ |\n|C3        |integer   |20      |2780     |2800 |4.3   |1.29  |1  |4   |5   |5   |6    |▁▂▁▂▆▁▇▅ |\n|C4        |integer   |26      |2774     |2800 |2.55  |1.38  |1  |1   |2   |4   |6    |▇▇▁▅▅▁▂▁ |\n|C5        |integer   |16      |2784     |2800 |3.3   |1.63  |1  |2   |3   |5   |6    |▆▇▁▅▇▁▆▃ |\n|E1        |integer   |23      |2777     |2800 |2.97  |1.63  |1  |2   |3   |4   |6    |▇▇▁▅▆▁▅▃ |\n|E2        |integer   |16      |2784     |2800 |3.14  |1.61  |1  |2   |3   |4   |6    |▆▇▁▅▇▁▅▃ |\n|E3        |integer   |25      |2775     |2800 |4     |1.35  |1  |3   |4   |5   |6    |▂▃▁▃▇▁▇▃ |\n|E4        |integer   |9       |2791     |2800 |4.42  |1.46  |1  |4   |5   |6   |6    |▁▂▁▂▃▁▇▆ |\n|E5        |integer   |21      |2779     |2800 |4.42  |1.33  |1  |4   |5   |5   |6    |▁▂▁▂▅▁▇▅ |\n|N1        |integer   |22      |2778     |2800 |2.93  |1.57  |1  |2   |3   |4   |6    |▇▇▁▅▆▁▅▂ |\n|N2        |integer   |21      |2779     |2800 |3.51  |1.53  |1  |2   |4   |5   |6    |▃▆▁▅▇▁▆▃ |\n|N3        |integer   |11      |2789     |2800 |3.22  |1.6   |1  |2   |3   |4   |6    |▆▇▁▅▇▁▆▃ |\n|N4        |integer   |36      |2764     |2800 |3.19  |1.57  |1  |2   |3   |4   |6    |▆▇▁▅▇▁▅▃ |\n|N5        |integer   |29      |2771     |2800 |2.97  |1.62  |1  |2   |3   |4   |6    |▇▇▁▅▆▁▃▃ |\n|O1        |integer   |22      |2778     |2800 |4.82  |1.13  |1  |4   |5   |6   |6    |▁▁▁▂▅▁▇▇ |\n|O2        |integer   |0       |2800     |2800 |2.71  |1.57  |1  |1   |2   |4   |6    |▇▇▁▃▅▁▃▂ |\n|O3        |integer   |28      |2772     |2800 |4.44  |1.22  |1  |4   |5   |5   |6    |▁▁▁▂▆▁▇▅ |\n|O4        |integer   |14      |2786     |2800 |4.89  |1.22  |1  |4   |5   |6   |6    |▁▁▁▁▃▁▆▇ |\n|O5        |integer   |20      |2780     |2800 |2.49  |1.33  |1  |1   |2   |3   |6    |▇▇▁▅▃▁▂▁ |\n|gender    |integer   |0       |2800     |2800 |1.67  |0.47  |1  |1   |2   |2   |2    |▃▁▁▁▁▁▁▇ |\n|education |integer   |223     |2577     |2800 |3.19  |1.11  |1  |3   |3   |4   |5    |▂▂▁▇▁▂▁▃ |\n|age       |integer   |0       |2800     |2800 |28.78 |11.13 |3  |20  |26  |35  |86   |▁▇▆▃▂▁▁▁ |\n\n### Note\nThis dataset was automatically described using the [codebook R package](https://rubenarslan.github.io/codebook/) (version 0.7.6.9000)."
#> }`
#> ```
#> </details>